@@ -1,6 +1,11 @@
 module github.com/v2rayA/v2rayA
 
-go 1.16
+go 1.21.0
+
+require (
+	github.com/daeuniverse/outbound v0.0.0-20230819151251-3ce60883d337
+	github.com/daeuniverse/softwind v0.0.0-20230903121035-afc8c5d27a4c
+)
 
 require (
 	github.com/adrg/xdg v0.4.0
