@@ -3,7 +3,11 @@
     <b-navbar ref="navs" fixed-top shadow type="is-light">
       <template slot="brand">
         <b-navbar-item href="/">
-          <img src="@/assets/img/logo2.png" alt="v2rayA" class="logo no-select" />
+          <img
+            src="@/assets/img/logo2.png"
+            alt="v2rayA"
+            class="logo no-select"
+          />
         </b-navbar-item>
         <b-navbar-item tag="div">
           <b-tag
@@ -85,14 +89,14 @@
         <b-dropdown
           position="is-bottom-left"
           aria-role="menu"
-          style="margin-right:10px"
+          style="margin-right: 10px"
           class="menudropdown"
         >
           <a slot="trigger" class="navbar-item" role="button">
             <span class="no-select">{{ username }}</span>
             <i
               class="iconfont icon-caret-down"
-              style="position: relative; top: 1px; left:2px"
+              style="position: relative; top: 1px; left: 2px"
             ></i>
           </a>
           <b-dropdown-item
@@ -105,14 +109,19 @@
             custom
             aria-role="menuitem"
             class="is-flex"
-            style="box-sizing: content-box;height: 16px;width: 60px;justify-content: space-between;"
+            style="
+              box-sizing: content-box;
+              height: 16px;
+              width: 60px;
+              justify-content: space-between;
+            "
           >
             <img
               v-for="lang of langs"
               :key="lang.flag"
               :src="require(`@/assets/img/flags/flag_${lang.flag}.svg`)"
               :alt="lang.alt"
-              style="height:100%;flex-shrink: 0;cursor: pointer"
+              style="height: 100%; flex-shrink: 0; cursor: pointer"
               @click="handleClickLang(lang.flag)"
             />
           </b-dropdown-item>
@@ -125,7 +134,7 @@
           >
             <i
               class="iconfont icon-logout"
-              style="position: relative;top:1px;"
+              style="position: relative; top: 1px"
             ></i>
             {{ $t("operations.logout") }}
           </b-dropdown-item>
@@ -172,23 +181,24 @@ export default {
       statusMap: {
         [this.$t("common.checkRunning")]: "is-light",
         [this.$t("common.notRunning")]: "is-danger",
-        [this.$t("common.isRunning")]: "is-success"
+        [this.$t("common.isRunning")]: "is-success",
       },
       coverStatusText: "",
       runningState: {
         running: this.$t("common.checkRunning"),
         connectedServer: null,
-        outboundToServerName: {}
+        outboundToServerName: {},
       },
       showCustomPorts: false,
       langs: [
         { flag: "zh", alt: "简体中文" },
-        { flag: "en", alt: "English" }
+        { flag: "en", alt: "English" },
+        { flag: "fa", alt: "فارسی" }
       ],
       outboundName: "proxy",
       outbounds: ["proxy"],
       outboundDropdownHover: {},
-      updateOutboundDropdown: true
+      updateOutboundDropdown: true,
     };
   },
   computed: {
@@ -202,7 +212,7 @@ export default {
     },
     isMobile() {
       return window.screen.width < 800;
-    }
+    },
   },
   mounted() {
     console.log("app created");
@@ -212,8 +222,8 @@ export default {
       document.title = `v2rayA - ${u.host}:${u.port}`;
     }
     this.$axios({
-      url: apiRoot + "/version"
-    }).then(res => {
+      url: apiRoot + "/version",
+    }).then((res) => {
       if (res.data.code === "SUCCESS") {
         let toastConf = {
           message: this.$t(
@@ -223,14 +233,14 @@ export default {
           type: "is-dark",
           position: "is-top",
           duration: 3000,
-          queue: false
+          queue: false,
         };
         if (res.data.data.foundNew) {
           toastConf.duration = 5000;
           toastConf.message +=
             ". " +
             this.$t("welcome.newVersion", {
-              version: res.data.data.remoteVersion
+              version: res.data.data.remoteVersion,
             });
           toastConf.type = "is-success";
         }
@@ -243,7 +253,7 @@ export default {
             type: "is-danger",
             position: "is-top",
             queue: false,
-            duration: 10000
+            duration: 10000,
           });
         } else if (!res.data.data.v5) {
           this.$buefy.toast.open({
@@ -251,15 +261,17 @@ export default {
             type: "is-danger",
             position: "is-top",
             queue: false,
-            duration: 10000
+            duration: 10000,
           });
         }
         localStorage["lite"] = res.data.data.lite;
+        localStorage["loadBalanceValid"] = res.data.data.loadBalanceValid;
+        localStorage["variant"] = res.data.data.variant;
       }
     });
     this.$axios({
-      url: apiRoot + "/outbounds"
-    }).then(res => {
+      url: apiRoot + "/outbounds",
+    }).then((res) => {
       if (res.data.code === "SUCCESS") {
         this.outbounds = res.data.data.outbounds;
       }
@@ -295,7 +307,7 @@ export default {
         // console.log("ws opened");
         //
       };
-      ws.onmessage = msg => {
+      ws.onmessage = (msg) => {
         msg.data && that.handleMessage(JSON.parse(msg.data));
       };
       ws.onclose = () => {
@@ -342,29 +354,29 @@ export default {
       this.$buefy.dialog.prompt({
         message: this.$t("outbound.addMessage"),
         inputAttrs: {
-          maxlength: 10
+          maxlength: 10,
         },
         trapFocus: true,
-        onConfirm: outbound => {
+        onConfirm: (outbound) => {
           let cancel;
           waitingConnected(
             this.$axios({
               url: apiRoot + "/outbound",
               method: "post",
               data: {
-                outbound
+                outbound,
               },
               cancelToken: new axios.CancelToken(function executor(c) {
                 cancel = c;
-              })
-            }).then(res => {
+              }),
+            }).then((res) => {
               if (res.data.code === "SUCCESS") {
                 this.$buefy.toast.open({
                   message: this.$t("common.success"),
                   type: "is-success",
                   duration: 2000,
                   position: "is-top",
-                  queue: false
+                  queue: false,
                 });
                 this.outbounds = res.data.data.outbounds;
               } else {
@@ -373,14 +385,14 @@ export default {
                   type: "is-warning",
                   duration: 5000,
                   position: "is-top",
-                  queue: false
+                  queue: false,
                 });
               }
             }),
             3 * 1000,
             cancel
           );
-        }
+        },
       });
     },
     handleDeleteOutbound(outbound) {
@@ -390,28 +402,29 @@ export default {
           url: apiRoot + "/outbound",
           method: "delete",
           data: {
-            outbound
+            outbound,
           },
           cancelToken: new axios.CancelToken(function executor(c) {
             cancel = c;
-          })
-        }).then(res => {
+          }),
+        }).then((res) => {
           if (res.data.code === "SUCCESS") {
             this.$buefy.toast.open({
               message: this.$t("common.success"),
               type: "is-success",
               duration: 2000,
               position: "is-top",
-              queue: false
+              queue: false,
             });
             this.outbounds = res.data.data.outbounds;
             if (this.outboundName === outbound) {
               this.outboundName = "proxy";
             }
             if (outbound in this.runningState.outboundToServerName) {
-              this.runningState.connectedServer = this.runningState.connectedServer.filter(
-                cs => cs.outbound !== outbound
-              );
+              this.runningState.connectedServer =
+                this.runningState.connectedServer.filter(
+                  (cs) => cs.outbound !== outbound
+                );
             }
           } else {
             this.$buefy.toast.open({
@@ -419,7 +432,7 @@ export default {
               type: "is-warning",
               duration: 5000,
               position: "is-top",
-              queue: false
+              queue: false,
             });
           }
         }),
@@ -436,13 +449,13 @@ export default {
         hasModalCard: true,
         canCancel: true,
         props: {
-          outbound: outbound
+          outbound: outbound,
         },
         events: {
           delete() {
             that.handleDeleteOutbound(outbound);
-          }
-        }
+          },
+        },
       });
     },
     handleClickLang(lang) {
@@ -475,8 +488,8 @@ export default {
         events: {
           clickPorts() {
             that.showCustomPorts = true;
-          }
-        }
+          },
+        },
       });
     },
     handleClickAbout() {
@@ -498,7 +511,7 @@ export default {
                         </a>
                     </footer>
                 </div>
-`
+`,
       });
     },
     handleClickStatus() {
@@ -510,12 +523,12 @@ export default {
             method: "post",
             cancelToken: new axios.CancelToken(function executor(c) {
               cancel = c;
-            })
-          }).then(res => {
+            }),
+          }).then((res) => {
             if (res.data.code === "SUCCESS") {
               Object.assign(this.runningState, {
                 running: this.$t("common.isRunning"),
-                connectedServer: res.data.data.touch.connectedServer
+                connectedServer: res.data.data.touch.connectedServer,
               });
             } else {
               this.$buefy.toast.open({
@@ -523,7 +536,7 @@ export default {
                 type: "is-warning",
                 duration: 5000,
                 position: "is-top",
-                queue: false
+                queue: false,
               });
             }
           }),
@@ -533,12 +546,12 @@ export default {
       } else if (this.runningState.running === this.$t("common.isRunning")) {
         this.$axios({
           url: apiRoot + "/v2ray",
-          method: "delete"
-        }).then(res => {
+          method: "delete",
+        }).then((res) => {
           if (res.data.code === "SUCCESS") {
             Object.assign(this.runningState, {
               running: this.$t("common.notRunning"),
-              connectedServer: res.data.data.touch.connectedServer
+              connectedServer: res.data.data.touch.connectedServer,
             });
           } else {
             this.$buefy.toast.open({
@@ -546,7 +559,7 @@ export default {
               type: "is-warning",
               duration: 5000,
               position: "is-top",
-              queue: false
+              queue: false,
             });
           }
         });
@@ -561,10 +574,10 @@ export default {
         parent: this,
         component: ModalLog,
         hasModalCard: true,
-        canCancel: true
+        canCancel: true,
       });
-    }
-  }
+    },
+  },
 };
 </script>
 
@@ -707,5 +720,10 @@ a {
   right: -1.5rem;
   top: 0;
   font-size: 1rem;
+}
+
+.navbar-item .dropdown-menu .dropdown-content {
+  max-height: calc(100vh - 60px);
+  overflow-y: auto;
 }
 </style>
