@@ -2,162 +2,162 @@
 export default [
     {
         "name": "Link_disconnect",
-        "unicode": "ea01",
+        "unicode": "554a",
         "svg": "<svg t=\"1577169047185\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"4374\" width=\"200\" height=\"200\"><path d=\"M484.96 523.328l-12.224 26.752-32.192 15.872 4.384 15.2c-20.512 32.416-22.656 82.912 0.352 116.16l91.68 133.088c34.368 49.984 97.888 61.696 147.84 27.328 43.136-35.52 54.912-98.816 27.456-147.264l-92.992-126.112c-28.704-41.504-76.672-57.664-115.744-43.2l-18.56-17.824z m93.888 82.592l91.776 133.088c17.216 24.96 10.72 59.968-14.368 77.152a56.096 56.096 0 0 1-77.472-14.368l-93.152-126.016c-11.36-16.512-14.656-38.976-4.768-51.776l22.912 33.216c11.328 16.768 32.48 20.736 42.144 7.776 15.488-4.288 19.392-25.408 7.808-41.952l-21.632-40.224c14.176 2.688 35.168 6.56 46.752 23.104zM379.872 148.416a105.088 105.088 0 0 0-81.12 122.656l21.824 156.576c13.472 45.888 44.672 73.568 86.752 81.344l8.992 30.656 36.416-0.448 25.216-17.152-3.008-22.496c40.352-21.504 63.488-68.064 57.344-112.672l-28.928-157.664c-12.032-53.024-63.136-91.456-123.488-80.8z m77.248 93.856l28.768 157.984c4.736 15.232-6.528 34.976-23.072 46.304l-1.76-29.344c-4.576-15.264-17.28-24.768-39.744-21.76-15.264 4.384-25.088 17.184-28.992 38.048l10.24 23.712c-19.872-10.88-25.568-19.136-28.608-41.504l-28.96-157.632c-8.96-30.656 17.44-54.88 48.352-63.776 29.44-1.76 55.04 17.344 63.776 47.968z m30.496 266.848l36-36.704 60.672 25.6-2.624 14.112-53.696-24.576-36.032 37.056-4.32-15.488z m-90.048 92.256l-54.528 18.944 2.624-13.952 46.016-13.248 5.248-28.032 30.784-8.96 5.856 8.544-23.776 10.016-12.224 26.688z\" p-id=\"4375\"></path></svg>"
     },
     {
         "name": "alert",
-        "unicode": "ea02",
+        "unicode": "554b",
         "svg": "<svg t=\"1577168856729\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2034\" width=\"200\" height=\"200\"><path d=\"M512 1024A512 512 0 1 1 512 0a512 512 0 0 1 0 1024zM448 192v384h128V192H448z m0 512v128h128v-128H448z\" fill=\"#FC746C\" p-id=\"2035\"></path></svg>"
     },
     {
         "name": "camera",
-        "unicode": "ea03",
+        "unicode": "554c",
         "svg": "<svg t=\"1603130360034\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"18939\" width=\"200\" height=\"200\"><path d=\"M864 260H728l-32.4-90.8C691 156.5 679 148 665.4 148H358.6c-13.5 0-25.6 8.5-30.1 21.2L296 260H160c-44.2 0-80 35.8-80 80v456c0 44.2 35.8 80 80 80h704c44.2 0 80-35.8 80-80V340c0-44.2-35.8-80-80-80zM512 716c-88.4 0-160-71.6-160-160s71.6-160 160-160 160 71.6 160 160-71.6 160-160 160z\" p-id=\"18940\"></path><path d=\"M512 556m-96 0a96 96 0 1 0 192 0 96 96 0 1 0-192 0Z\" p-id=\"18941\"></path></svg>"
     },
     {
         "name": "caret-down",
-        "unicode": "ea04",
+        "unicode": "554d",
         "svg": "<svg t=\"1577169193557\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5195\" width=\"200\" height=\"200\"><path d=\"M840.4 300H183.6c-19.7 0-30.7 20.8-18.5 35l328.4 380.8c9.4 10.9 27.5 10.9 37 0L858.9 335c12.2-14.2 1.2-35-18.5-35z\" p-id=\"5196\"></path></svg>"
     },
     {
         "name": "caret-up",
-        "unicode": "ea05",
+        "unicode": "554e",
         "svg": "<svg t=\"1577168830461\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"1812\" width=\"200\" height=\"200\"><path d=\"M858.9 689L530.5 308.2c-9.4-10.9-27.5-10.9-37 0L165.1 689c-12.2 14.2-1.2 35 18.5 35h656.8c19.7 0 30.7-20.8 18.5-35z\" p-id=\"1813\"></path></svg>"
     },
     {
         "name": "chakandaorujilu",
-        "unicode": "ea06",
+        "unicode": "554f",
         "svg": "<svg t=\"1577169003222\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"3546\" width=\"200\" height=\"200\"><path d=\"M896 192l-64-64-320 320-96-96v256h256l-96-96 320-320z\" p-id=\"3547\"></path><path d=\"M832 831.89l-0.11 0.11H192.11l-0.11-0.11V192.11l0.11-0.11H704v-64H192a64.19 64.19 0 0 0-64 64v640a64.19 64.19 0 0 0 64 64h640a64.19 64.19 0 0 0 64-64V320h-64z\" p-id=\"3548\"></path></svg>"
     },
     {
         "name": "check-circle-fill",
-        "unicode": "ea07",
+        "unicode": "5550",
         "svg": "<svg t=\"1627629994056\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2816\" width=\"200\" height=\"200\"><path d=\"M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4 64 512 64z m193.5 301.7l-210.6 292c-12.7 17.7-39 17.7-51.7 0L318.5 484.9c-3.8-5.3 0-12.7 6.5-12.7h46.9c10.2 0 19.9 4.9 25.9 13.3l71.2 98.8 157.2-218c6-8.3 15.6-13.3 25.9-13.3H699c6.5 0 10.3 7.4 6.5 12.7z\" p-id=\"2817\"></path></svg>"
     },
     {
         "name": "chuangjiangongdan1",
-        "unicode": "ea08",
+        "unicode": "5551",
         "svg": "<svg t=\"1577169032634\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"4021\" width=\"200\" height=\"200\"><path d=\"M704.137 127.99V63.946H320.062v64.044H127.689l-0.83 832.355 321.228-0.787v-63.919h-256.05l-0.336-703.856h128.361l-0.097 64.169h384.172v-64.169l128.025 0.055v384.207h64.012v-448.22l-192.037 0.165z m-64.013 63.794h-256.05V127.99h256.05v63.794z\" p-id=\"4022\"></path><path d=\"M320.062 511.014l256.05 0.58v64.452h-256.05zM320.062 383.79h384.075v64.419H320.062zM768.149 768.149l128.025-0.347v63.919H768.149v127.837h-64.012V831.721H576.112v-63.919h128.025V639.965h64.012z\" p-id=\"4023\"></path><path d=\"M320.062 639.965l128.025-0.534v64.452l-128.025 1.114z\" p-id=\"4024\"></path></svg>"
     },
     {
         "name": "clock2",
-        "unicode": "ea09",
+        "unicode": "5552",
         "svg": "<svg t=\"1577168911222\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2604\" width=\"200\" height=\"200\"><path d=\"M512 0c-282.784 0-512 229.216-512 512s229.216 512 512 512 512-229.216 512-512-229.216-512-512-512zM658.752 749.248l-210.752-210.752 0-282.496 128 0 0 229.504 173.248 173.248-90.528 90.528z\" p-id=\"2605\"></path></svg>"
     },
     {
         "name": "close-circle-fill",
-        "unicode": "ea0a",
+        "unicode": "5553",
         "svg": "<svg t=\"1627635178003\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"12383\" width=\"200\" height=\"200\"><path d=\"M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4 64 512 64z m165.4 618.2l-66-0.3L512 563.4l-99.3 118.4-66.1 0.3c-4.4 0-8-3.5-8-8 0-1.9 0.7-3.7 1.9-5.2l130.1-155L340.5 359c-1.2-1.5-1.9-3.3-1.9-5.2 0-4.4 3.6-8 8-8l66.1 0.3L512 464.6l99.3-118.4 66-0.3c4.4 0 8 3.5 8 8 0 1.9-0.7 3.7-1.9 5.2L553.5 514l130 155c1.2 1.5 1.9 3.3 1.9 5.2 0 4.4-3.6 8-8 8z\" p-id=\"12384\"></path></svg>"
     },
     {
         "name": "cloud-sync",
-        "unicode": "ea0b",
+        "unicode": "5554",
         "svg": "<svg t=\"1577169221697\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5547\" width=\"200\" height=\"200\"><path d=\"M811.4 368.9C765.6 248 648.9 162 512.2 162S258.8 247.9 213 368.8C126.9 391.5 63.5 470.2 64 563.6 64.6 668 145.6 752.9 247.6 762c4.7 0.4 8.7-3.3 8.7-8v-60.4c0-4-3-7.4-7-7.9-27-3.4-52.5-15.2-72.1-34.5-24-23.5-37.2-55.1-37.2-88.6 0-28 9.1-54.4 26.2-76.4 16.7-21.4 40.2-36.9 66.1-43.7l37.9-10 13.9-36.7c8.6-22.8 20.6-44.2 35.7-63.5 14.9-19.2 32.6-36 52.4-50 41.1-28.9 89.5-44.2 140-44.2s98.9 15.3 140 44.3c19.9 14 37.5 30.8 52.4 50 15.1 19.3 27.1 40.7 35.7 63.5l13.8 36.6 37.8 10c54.2 14.4 92.1 63.7 92.1 120 0 33.6-13.2 65.1-37.2 88.6-19.5 19.2-44.9 31.1-71.9 34.5-4 0.5-6.9 3.9-6.9 7.9V754c0 4.7 4.1 8.4 8.8 8 101.7-9.2 182.5-94 183.2-198.2 0.6-93.4-62.7-172.1-148.6-194.9z\" p-id=\"5548\"></path><path d=\"M376.9 656.4c1.8-33.5 15.7-64.7 39.5-88.6 25.4-25.5 60-39.8 96-39.8 36.2 0 70.3 14.1 96 39.8 1.4 1.4 2.7 2.8 4.1 4.3l-25 19.6c-5.3 4.1-3.5 12.5 3 14.1l98.2 24c5 1.2 9.9-2.6 9.9-7.7l0.5-101.3c0-6.7-7.6-10.5-12.9-6.3L663 532.7c-36.6-42-90.4-68.6-150.5-68.6-107.4 0-195 85.1-199.4 191.7-0.2 4.5 3.4 8.3 8 8.3H369c4.2-0.1 7.7-3.4 7.9-7.7zM703 664h-47.9c-4.2 0-7.7 3.3-8 7.6-1.8 33.5-15.7 64.7-39.5 88.6-25.4 25.5-60 39.8-96 39.8-36.2 0-70.3-14.1-96-39.8-1.4-1.4-2.7-2.8-4.1-4.3l25-19.6c5.3-4.1 3.5-12.5-3-14.1l-98.2-24c-5-1.2-9.9 2.6-9.9 7.7l-0.4 101.4c0 6.7 7.6 10.5 12.9 6.3l23.2-18.2c36.6 42 90.4 68.6 150.5 68.6 107.4 0 195-85.1 199.4-191.7 0.2-4.5-3.4-8.3-8-8.3z\" p-id=\"5549\"></path></svg>"
     },
     {
         "name": "cloud",
-        "unicode": "ea0c",
+        "unicode": "5555",
         "svg": "<svg t=\"1577169038912\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"4257\" width=\"200\" height=\"200\"><path d=\"M811.4 418.7C765.6 297.9 648.9 212 512.2 212S258.8 297.8 213 418.6C127.3 441.1 64 519.1 64 612c0 110.5 89.5 200 199.9 200h496.2C870.5 812 960 722.5 960 612c0-92.7-63.1-170.7-148.6-193.3z m36.3 281c-23.4 23.4-54.5 36.3-87.6 36.3H263.9c-33.1 0-64.2-12.9-87.6-36.3-23.4-23.4-36.3-54.6-36.3-87.7 0-28 9.1-54.3 26.2-76.3 16.7-21.3 40.2-36.8 66.1-43.7l37.9-9.9 13.9-36.6c8.6-22.8 20.6-44.1 35.7-63.4 14.9-19.2 32.6-35.9 52.4-49.9 41.1-28.9 89.5-44.2 140-44.2s98.9 15.3 140 44.2c19.9 14 37.5 30.8 52.4 49.9 15.1 19.3 27.1 40.7 35.7 63.4l13.8 36.5 37.8 10c54.3 14.5 92.1 63.8 92.1 120 0 33.1-12.9 64.3-36.3 87.7z\" p-id=\"4258\"></path></svg>"
     },
     {
         "name": "coding_alert_circle",
-        "unicode": "ea0d",
+        "unicode": "5556",
         "svg": "<svg t=\"1577168867373\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2145\" width=\"200\" height=\"200\"><path d=\"M512 1024C229.696 1024 0 794.304 0 512S229.696 0 512 0s512 229.696 512 512-229.696 512-512 512z m0-960C264.96 64 64 264.96 64 512s200.96 448 448 448 448-200.96 448-448S759.04 64 512 64z\" fill=\"#FD5C3F\" p-id=\"2146\"></path><path d=\"M480 256h64v384h-64zM480 704h64v64h-64z\" fill=\"#FD5C3F\" p-id=\"2147\"></path></svg>"
     },
     {
         "name": "daoru",
-        "unicode": "ea0e",
+        "unicode": "5557",
         "svg": "<svg t=\"1577169019177\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"3783\" width=\"200\" height=\"200\"><path d=\"M111.6 148.8c7.9 0 16.9 17.9 21.6 32.2 25.9 88.7 143 182.6 248.6 194.2v-1.1c0-29.6 10.6-50.7 31.7-64.4 11.1-6.9 22.7-10.6 34.8-10.6 13.7 0 27.4 4.8 39.1 13.7L645.8 438c17.9 13.2 28.5 34.8 28.5 57.5 0 22.2-10.6 43.8-28 57L487.4 678.2c-1.6 1.1-2.6 2.1-4.2 2.6-7.4 4.2-19.5 9.5-34.3 9.5-12.1 0-23.7-3.2-34.3-10-26.4-16.9-30.6-43.8-32.2-55.4l-0.5-2.1c-0.5-1.6-0.5-3.7-0.5-5.3v-5.3c-162.6-26.9-314.1-192.6-314.1-352 0-29.6 4.2-59.6 12.7-88.1 4.2-13.8 18.4-26.5 31.6-23.3z m335.2 467.1c0.5 2.6 1.1 6.9 2.1 8.4l156.2-123.5c1.6-1.1 2.6-3.2 2.6-5.3 0-2.1-0.5-3.7-2.6-5.3L447.8 365.7c0 1.1-0.5 3.2-0.5 7.9v34.8c0 17.4-13.2 31.7-30.6 32.7-104.5 7.4-213.8-61.2-280.3-145.7 22.2 123 154.1 247.5 279.7 254.4 17.4 1.1 31.1 15.3 31.1 32.7v32.7l-0.4 0.7z m0.5-0.5\" p-id=\"3784\"></path><path d=\"M309 916.3c-67 0-121.4-54.9-121.4-121.4V593.2h69.1v201.6c0 29 23.8 52.3 52.3 52.3h527.3c29 0 52.3-23.8 52.3-52.3V226.4c0-29-23.8-52.3-52.3-52.3H399.8v-69.7h436.5c67 0 121.9 54.9 121.9 121.9v567.9c0 67-54.9 121.9-121.9 121.9H309v0.2z m0 0\" p-id=\"3785\"></path><path d=\"M841.6 626H729.1v-59.1h112.4V626z m0 154.1H545.5V721h296.1v59.1z m0-425.4h-237v-59.1h237v59.1z m0 140.9H729.1v-59.1h112.4v59.1z m0-59.1\" p-id=\"3786\"></path></svg>"
     },
     {
         "name": "daoruzupu-xianxing",
-        "unicode": "ea0f",
+        "unicode": "5558",
         "svg": "<svg t=\"1577168978161\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"3310\" width=\"200\" height=\"200\"><path d=\"M888.3 757.4h-53.8c-4.2 0-7.7 3.5-7.7 7.7v61.8H197.1V197.1h629.8v61.8c0 4.2 3.5 7.7 7.7 7.7h53.8c4.2 0 7.7-3.4 7.7-7.7V158.7c0-17-13.7-30.7-30.7-30.7H158.7c-17 0-30.7 13.7-30.7 30.7v706.6c0 17 13.7 30.7 30.7 30.7h706.6c17 0 30.7-13.7 30.7-30.7V765.1c0-4.3-3.5-7.7-7.7-7.7z\" p-id=\"3311\"></path><path d=\"M902 476H588v-76c0-6.7-7.8-10.5-13-6.3l-141.9 112c-4.1 3.2-4.1 9.4 0 12.6l141.9 112c5.3 4.2 13 0.4 13-6.3v-76h314c4.4 0 8-3.6 8-8v-56c0-4.4-3.6-8-8-8z\" p-id=\"3312\"></path></svg>"
     },
     {
         "name": "delete",
-        "unicode": "ea10",
+        "unicode": "5559",
         "svg": "<svg t=\"1577169183195\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5077\" width=\"200\" height=\"200\"><path d=\"M360 184h-8c4.4 0 8-3.6 8-8v8h304v-8c0 4.4 3.6 8 8 8h-8v72h72v-80c0-35.3-28.7-64-64-64H352c-35.3 0-64 28.7-64 64v80h72v-72z\" p-id=\"5078\"></path><path d=\"M864 256H160c-17.7 0-32 14.3-32 32v32c0 4.4 3.6 8 8 8h60.4l24.7 523c1.6 34.1 29.8 61 63.9 61h454c34.2 0 62.3-26.8 63.9-61l24.7-523H888c4.4 0 8-3.6 8-8v-32c0-17.7-14.3-32-32-32zM731.3 840H292.7l-24.2-512h487l-24.2 512z\" p-id=\"5079\"></path></svg>"
     },
     {
         "name": "dian",
-        "unicode": "ea11",
+        "unicode": "555a",
         "svg": "<svg t=\"1577168983543\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"3428\" width=\"200\" height=\"200\"><path d=\"M810.7 549.3v336H213.3V138.7h261.3c20.6 0 37.3-16.7 37.3-37.3 0-20.6-16.7-37.3-37.3-37.3H213.3c-41.2 0-74.7 33.4-74.7 74.7v746.7c0 41.2 33.4 74.7 74.7 74.7h597.3c41.2 0 74.7-33.4 74.7-74.7v-336c0-20.6-16.7-37.3-37.3-37.3-20.6-0.2-37.3 16.5-37.3 37.1z\" p-id=\"3429\"></path><path d=\"M874.3 75c-14.7-14.7-38.4-14.7-53.1 0L624.4 271.8v-59.5c0-20.7-16.8-37.5-37.5-37.5s-37.5 16.8-37.5 37.5v150.2c0 20.7 16.8 37.5 37.5 37.5H737c20.7 0 37.5-16.8 37.5-37.5S757.7 325 737 325h-59.5l196.8-196.8c14.7-14.8 14.7-38.5 0-53.2z\" p-id=\"3430\"></path></svg>"
     },
     {
         "name": "earth",
-        "unicode": "ea12",
+        "unicode": "555b",
         "svg": "<svg t=\"1577169239665\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5783\" width=\"200\" height=\"200\"><path d=\"M854.4 800.9c0.2-0.3 0.5-0.6 0.7-0.9C920.6 722.1 960 621.7 960 512s-39.4-210.1-104.8-288c-0.2-0.3-0.5-0.5-0.7-0.8-1.1-1.3-2.1-2.5-3.2-3.7-0.4-0.5-0.8-0.9-1.2-1.4-1.4-1.6-2.7-3.1-4.1-4.7l-0.1-0.1c-1.5-1.7-3.1-3.4-4.6-5.1l-0.1-0.1c-3.2-3.4-6.4-6.8-9.7-10.1l-0.1-0.1-4.8-4.8-0.3-0.3c-1.5-1.5-3-2.9-4.5-4.3-0.5-0.5-1-1-1.6-1.5-1-1-2-1.9-3-2.8-0.3-0.3-0.7-0.6-1-1C736.4 109.2 629.5 64 512 64s-224.4 45.2-304.3 119.2c-0.3 0.3-0.7 0.6-1 1-1 0.9-2 1.9-3 2.9-0.5 0.5-1 1-1.6 1.5-1.5 1.4-3 2.9-4.5 4.3l-0.3 0.3-4.8 4.8-0.1 0.1c-3.3 3.3-6.5 6.7-9.7 10.1l-0.1 0.1c-1.6 1.7-3.1 3.4-4.6 5.1l-0.1 0.1c-1.4 1.5-2.8 3.1-4.1 4.7-0.4 0.5-0.8 0.9-1.2 1.4-1.1 1.2-2.1 2.5-3.2 3.7-0.2 0.3-0.5 0.5-0.7 0.8C103.4 301.9 64 402.3 64 512s39.4 210.1 104.8 288c0.2 0.3 0.5 0.6 0.7 0.9 1 1.2 2.1 2.5 3.1 3.7 0.4 0.5 0.8 0.9 1.2 1.4 1.4 1.6 2.7 3.1 4.1 4.7 0 0.1 0.1 0.1 0.1 0.2 1.5 1.7 3 3.4 4.6 5l0.1 0.1c3.2 3.4 6.4 6.8 9.6 10.1l0.1 0.1c1.6 1.6 3.1 3.2 4.7 4.7l0.3 0.3c3.3 3.3 6.7 6.5 10.1 9.6 80.1 74 187 119.2 304.5 119.2s224.4-45.2 304.3-119.2c3.4-3.1 6.7-6.3 10-9.6l0.3-0.3c1.6-1.6 3.2-3.1 4.7-4.7l0.1-0.1c3.3-3.3 6.5-6.7 9.6-10.1l0.1-0.1c1.5-1.7 3.1-3.3 4.6-5 0-0.1 0.1-0.1 0.1-0.2 1.4-1.5 2.8-3.1 4.1-4.7 0.4-0.5 0.8-0.9 1.2-1.4 1.2-1.3 2.3-2.5 3.3-3.7z m4.1-142.6c-13.8 32.6-32 62.8-54.2 90.2-24.9-21.5-52.2-40.3-81.5-55.9 11.6-46.9 18.8-98.4 20.7-152.6H887c-3 40.9-12.6 80.6-28.5 118.3zM887 484H743.5c-1.9-54.2-9.1-105.7-20.7-152.6 29.3-15.6 56.6-34.4 81.5-55.9 22.2 27.4 40.4 57.6 54.2 90.2C874.4 403.4 884 443.1 887 484zM658.3 165.5c39.7 16.8 75.8 40 107.6 69.2-18.5 15.8-38.4 29.7-59.4 41.8-15.7-45-35.8-84.1-59.2-115.4 3.7 1.4 7.4 2.9 11 4.4z m-90.6 700.6c-9.2 7.2-18.4 12.7-27.7 16.4V697c39.9 2.8 78.6 11.6 115.7 26.2-8.3 24.6-17.9 47.3-29 67.8-17.4 32.4-37.8 58.3-59 75.1z m59-633.1c11 20.6 20.7 43.3 29 67.8-37.1 14.6-75.8 23.4-115.7 26.2V141.6c9.2 3.7 18.5 9.1 27.7 16.4 21.2 16.7 41.6 42.6 59 75zM540 640.9V540h147.5c-1.6 44.2-7.1 87.1-16.3 127.8l-0.3 1.2c-41.1-15.6-85.1-25.3-130.9-28.1z m0-156.9V383.1c45.8-2.8 89.8-12.5 130.9-28.1l0.3 1.2c9.2 40.7 14.7 83.5 16.3 127.8H540z m-56 56v100.9c-45.8 2.8-89.8 12.5-130.9 28.1l-0.3-1.2c-9.2-40.7-14.7-83.5-16.3-127.8H484z m-147.5-56c1.6-44.2 7.1-87.1 16.3-127.8l0.3-1.2c41.1 15.6 85 25.3 130.9 28.1V484H336.5zM484 697v185.4c-9.2-3.7-18.5-9.1-27.7-16.4-21.2-16.7-41.7-42.7-59.1-75.1-11-20.6-20.7-43.3-29-67.8 37.2-14.6 75.9-23.3 115.8-26.1z m0-370c-39.9-2.8-78.6-11.6-115.7-26.2 8.3-24.6 17.9-47.3 29-67.8 17.4-32.4 37.8-58.4 59.1-75.1 9.2-7.2 18.4-12.7 27.7-16.4V327zM365.7 165.5c3.7-1.5 7.3-3 11-4.4-23.4 31.3-43.5 70.4-59.2 115.4-21-12-40.9-26-59.4-41.8 31.8-29.2 67.9-52.4 107.6-69.2zM165.5 365.7c13.8-32.6 32-62.8 54.2-90.2 24.9 21.5 52.2 40.3 81.5 55.9-11.6 46.9-18.8 98.4-20.7 152.6H137c3-40.9 12.6-80.6 28.5-118.3zM137 540h143.5c1.9 54.2 9.1 105.7 20.7 152.6-29.3 15.6-56.6 34.4-81.5 55.9-22.2-27.4-40.4-57.6-54.2-90.2C149.6 620.6 140 580.9 137 540z m228.7 318.5c-39.7-16.8-75.8-40-107.6-69.2 18.5-15.8 38.4-29.7 59.4-41.8 15.7 45 35.8 84.1 59.2 115.4-3.7-1.4-7.4-2.9-11-4.4z m292.6 0c-3.7 1.5-7.3 3-11 4.4 23.4-31.3 43.5-70.4 59.2-115.4 21 12 40.9 26 59.4 41.8-31.8 29.2-67.9 52.4-107.6 69.2z\" p-id=\"5784\"></path></svg>"
     },
     {
         "name": "heart",
-        "unicode": "ea13",
+        "unicode": "555c",
         "svg": "<svg t=\"1577169213881\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5430\" width=\"200\" height=\"200\"><path d=\"M923 283.6c-13.4-31.1-32.6-58.9-56.9-82.8-24.3-23.8-52.5-42.4-84-55.5-32.5-13.5-66.9-20.3-102.4-20.3-49.3 0-97.4 13.5-139.2 39-10 6.1-19.5 12.8-28.5 20.1-9-7.3-18.5-14-28.5-20.1-41.8-25.5-89.9-39-139.2-39-35.5 0-69.9 6.8-102.4 20.3-31.4 13-59.7 31.7-84 55.5-24.4 23.9-43.5 51.7-56.9 82.8-13.9 32.3-21 66.6-21 101.9 0 33.3 6.8 68 20.3 103.3 11.3 29.5 27.5 60.1 48.2 91 32.8 48.9 77.9 99.9 133.9 151.6 92.8 85.7 184.7 144.9 188.6 147.3l23.7 15.2c10.5 6.7 24 6.7 34.5 0l23.7-15.2c3.9-2.5 95.7-61.6 188.6-147.3 56-51.7 101.1-102.7 133.9-151.6 20.7-30.9 37-61.5 48.2-91 13.5-35.3 20.3-70 20.3-103.3 0.1-35.3-7-69.6-20.9-101.9zM512 814.8S156 586.7 156 385.5C156 283.6 240.3 201 344.3 201c73.1 0 136.5 40.8 167.7 100.4C543.2 241.8 606.6 201 679.7 201c104 0 188.3 82.6 188.3 184.5 0 201.2-356 429.3-356 429.3z\" p-id=\"5431\"></path></svg>"
     },
     {
         "name": "help-circle-outline",
-        "unicode": "ea14",
+        "unicode": "555d",
         "svg": "<svg t=\"1577168739988\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"1695\" width=\"200\" height=\"200\"><path d=\"M469.333333 768h85.333334v-85.333333h-85.333334v85.333333m42.666667-682.666667C276.48 85.333333 85.333333 276.48 85.333333 512s191.146667 426.666667 426.666667 426.666667 426.666667-191.146667 426.666667-426.666667S747.52 85.333333 512 85.333333m0 768c-188.16 0-341.333333-153.173333-341.333333-341.333333s153.173333-341.333333 341.333333-341.333333 341.333333 153.173333 341.333333 341.333333-153.173333 341.333333-341.333333 341.333333m0-597.333333c-94.293333 0-170.666667 76.373333-170.666667 170.666667h85.333334c0-46.933333 38.4-85.333333 85.333333-85.333334s85.333333 38.4 85.333333 85.333334c0 85.333333-128 74.666667-128 213.333333h85.333334c0-96 128-106.666667 128-213.333333 0-94.293333-76.373333-170.666667-170.666667-170.666667z\" p-id=\"1696\"></path></svg>"
     },
     {
         "name": "info",
-        "unicode": "ea15",
+        "unicode": "555e",
         "svg": "<svg style=\"width:24px;height:24px\" viewBox=\"0 0 24 24\">\n    <path fill=\"currentColor\"\n        d=\"M11,9H13V7H11M12,20C7.59,20 4,16.41 4,12C4,7.59 7.59,4 12,4C16.41,4 20,7.59 20,12C20,16.41 16.41,20 12,20M12,2A10,10 0 0,0 2,12A10,10 0 0,0 12,22A10,10 0 0,0 22,12A10,10 0 0,0 12,2M11,17H13V11H11V17Z\" />\n</svg>"
     },
     {
         "name": "label",
-        "unicode": "ea16",
+        "unicode": "555f",
         "svg": "<svg t=\"1577168845349\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"1923\" width=\"200\" height=\"200\"><path d=\"M752.213333 249.173333C736.853333 227.413333 711.253333 213.333333 682.666667 213.333333L213.333333 213.76C166.4 213.76 128 251.733333 128 298.666667v426.666666c0 46.933333 38.4 84.906667 85.333333 84.906667L682.666667 810.666667c28.586667 0 54.186667-14.08 69.546666-35.84L938.666667 512l-186.453334-262.826667z\" p-id=\"1924\"></path></svg>"
     },
     {
         "name": "lianjie",
-        "unicode": "ea17",
+        "unicode": "5560",
         "svg": "<svg t=\"1577169177516\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"4960\" width=\"200\" height=\"200\"><path d=\"M591.1 432.7c-9.7-9.7-25.5-9.7-35.2 0L432.6 555.9c-9.7 9.7-9.7 25.5 0 35.2 9.7 9.7 25.5 9.7 35.2 0l123.3-123.3c9.7-9.6 9.7-25.4 0-35.1z m-288.2 387c-19.4 19.5-54.4-5.2-73.9-24.6-19.5-19.4-44.1-54.4-24.6-73.9L431.6 486c-33.2-8.8-78.1-12.5-104.1 13.5L155.2 672c-38.9 38.9-34.4 117.6 4.5 156.5l35.2 35.2c38.9 38.9 118.4 44.2 157.3 5.3l172.4-172.4c26-26 21.5-71.7 12.7-104.9l-234.4 228zM864 194.6l-35.2-35.2c-38.9-38.9-117.6-43.4-156.5-4.5L499.9 327.2c-26 26-22.3 70.9-13.5 104.1l235.2-227.2c19.4-19.5 54.4 5.2 73.9 24.6 19.4 19.5 44.1 54.4 24.6 73.9L592 537c33.2 8.8 78.9 13.3 104.9-12.7l172.4-172.4c38.9-39 33.6-118.4-5.3-157.3z\" p-id=\"4961\"></path></svg>"
     },
     {
         "name": "loading_ico",
-        "unicode": "ea18",
+        "unicode": "5561",
         "svg": "<svg t=\"1577168893140\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2368\" width=\"200\" height=\"200\"><path d=\"M511.99999895 511.99924322C511.9995814 511.99924322 511.99924266 511.99958223 511.99924266 512.00000029 511.99924266 512.00041842 511.9995814 512.00075734 511.99999895 512.00075734a0.00075678 0.00075678 0 0 0 0.0007566-0.00075705C512.00075536 511.99958223 512.00041654 511.99924322 511.99999895 511.99924322z m0 0.00141151c-0.00036048 0-0.00065377-0.00029348-0.00065377-0.00065444 0-0.00036086 0.00029327-0.00065443 0.00065377-0.00065444 0.00036069 0 0.00065398 0.00029348 0.00065397 0.00065444-7e-8 0.00036097-0.00029327 0.00065443-0.00065397 0.00065444z\" p-id=\"2369\"></path><path d=\"M781.62850247 279.31175122c0-221.11127931-257.87478539-221.11127931-257.87478538-221.11127931v442.13418482c-0.08837381 0 257.87478539 0 257.87478538-221.02290551zM279.31175122 242.45987135c-221.11127931 0-221.11127931 257.87478539-221.11127931 257.87478538h442.13418482s0-257.87478539-221.02290551-257.87478538zM242.45987135 744.86499641c0 220.9345317 257.87478539 220.9345317 257.87478538 220.93453168V523.75371709s-257.87478539 0-257.87478538 221.11127932zM744.86499641 781.62850247c220.9345317 0 220.9345317-257.87478539 220.93453168-257.87478538H523.66534327s0 257.87478539 221.19965314 257.87478538z\" p-id=\"2370\"></path></svg>"
     },
     {
         "name": "logout",
-        "unicode": "ea19",
+        "unicode": "5562",
         "svg": "<svg t=\"1577169204223\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5312\" width=\"200\" height=\"200\"><path d=\"M868 732h-70.3c-4.8 0-9.3 2.1-12.3 5.8-7 8.5-14.5 16.7-22.4 24.5-32.6 32.5-70.5 58.1-112.7 75.9-43.6 18.4-90 27.8-137.9 27.8-47.9 0-94.3-9.4-137.9-27.8-42.2-17.8-80.1-43.4-112.7-75.9-32.6-32.5-58.1-70.4-76-112.5C167.3 606.2 158 559.9 158 512s9.4-94.2 27.8-137.8c17.8-42.1 43.4-80 76-112.5s70.5-58.1 112.7-75.9c43.6-18.4 90-27.8 137.9-27.8 47.9 0 94.3 9.3 137.9 27.8 42.2 17.8 80.1 43.4 112.7 75.9 7.9 7.9 15.3 16.1 22.4 24.5 3 3.7 7.6 5.8 12.3 5.8H868c6.3 0 10.2-7 6.7-12.3C798 160.5 663.8 81.6 511.3 82 271.7 82.6 79.6 277.1 82 516.4 84.4 751.9 276.2 942 512.4 942c152.1 0 285.7-78.8 362.3-197.7 3.4-5.3-0.4-12.3-6.7-12.3z\" p-id=\"5313\"></path><path d=\"M956.9 505.7L815 393.7c-5.3-4.2-13-0.4-13 6.3v76H488c-4.4 0-8 3.6-8 8v56c0 4.4 3.6 8 8 8h314v76c0 6.7 7.8 10.5 13 6.3l141.9-112c4.1-3.2 4.1-9.4 0-12.6z\" p-id=\"5314\"></path></svg>"
     },
     {
         "name": "setting",
-        "unicode": "ea1a",
+        "unicode": "5563",
         "svg": "<svg t=\"1577169232326\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"5665\" width=\"200\" height=\"200\"><path d=\"M924.8 625.7l-65.5-56c3.1-19 4.7-38.4 4.7-57.8s-1.6-38.8-4.7-57.8l65.5-56c10.1-8.6 13.8-22.6 9.3-35.2l-0.9-2.6c-18.1-50.5-44.9-96.9-79.7-137.9l-1.8-2.1c-8.6-10.1-22.5-13.9-35.1-9.5l-81.3 28.9c-30-24.6-63.5-44-99.7-57.6l-15.7-85c-2.4-13.1-12.7-23.3-25.8-25.7l-2.7-0.5c-52.1-9.4-106.9-9.4-159 0l-2.7 0.5c-13.1 2.4-23.4 12.6-25.8 25.7l-15.8 85.4c-35.9 13.6-69.2 32.9-99 57.4l-81.9-29.1c-12.5-4.4-26.5-0.7-35.1 9.5l-1.8 2.1c-34.8 41.1-61.6 87.5-79.7 137.9l-0.9 2.6c-4.5 12.5-0.8 26.5 9.3 35.2l66.3 56.6c-3.1 18.8-4.6 38-4.6 57.1 0 19.2 1.5 38.4 4.6 57.1L99 625.5c-10.1 8.6-13.8 22.6-9.3 35.2l0.9 2.6c18.1 50.4 44.9 96.9 79.7 137.9l1.8 2.1c8.6 10.1 22.5 13.9 35.1 9.5l81.9-29.1c29.8 24.5 63.1 43.9 99 57.4l15.8 85.4c2.4 13.1 12.7 23.3 25.8 25.7l2.7 0.5c26.1 4.7 52.8 7.1 79.5 7.1 26.7 0 53.5-2.4 79.5-7.1l2.7-0.5c13.1-2.4 23.4-12.6 25.8-25.7l15.7-85c36.2-13.6 69.7-32.9 99.7-57.6l81.3 28.9c12.5 4.4 26.5 0.7 35.1-9.5l1.8-2.1c34.8-41.1 61.6-87.5 79.7-137.9l0.9-2.6c4.5-12.3 0.8-26.3-9.3-35zM788.3 465.9c2.5 15.1 3.8 30.6 3.8 46.1s-1.3 31-3.8 46.1l-6.6 40.1 74.7 63.9c-11.3 26.1-25.6 50.7-42.6 73.6L721 702.8l-31.4 25.8c-23.9 19.6-50.5 35-79.3 45.8l-38.1 14.3-17.9 97c-28.1 3.2-56.8 3.2-85 0l-17.9-97.2-37.8-14.5c-28.5-10.8-55-26.2-78.7-45.7l-31.4-25.9-93.4 33.2c-17-22.9-31.2-47.6-42.6-73.6l75.5-64.5-6.5-40c-2.4-14.9-3.7-30.3-3.7-45.5 0-15.3 1.2-30.6 3.7-45.5l6.5-40-75.5-64.5c11.3-26.1 25.6-50.7 42.6-73.6l93.4 33.2 31.4-25.9c23.7-19.5 50.2-34.9 78.7-45.7l37.9-14.3 17.9-97.2c28.1-3.2 56.8-3.2 85 0l17.9 97 38.1 14.3c28.7 10.8 55.4 26.2 79.3 45.8l31.4 25.8 92.8-32.9c17 22.9 31.2 47.6 42.6 73.6L781.8 426l6.5 39.9z\" p-id=\"5666\"></path><path d=\"M512 326c-97.2 0-176 78.8-176 176s78.8 176 176 176 176-78.8 176-176-78.8-176-176-176z m79.2 255.2C570 602.3 541.9 614 512 614c-29.9 0-58-11.7-79.2-32.8C411.7 560 400 531.9 400 502c0-29.9 11.7-58 32.8-79.2C454 401.6 482.1 390 512 390c29.9 0 58 11.6 79.2 32.8C612.3 444 624 472.1 624 502c0 29.9-11.7 58-32.8 79.2z\" p-id=\"5667\"></path></svg>"
     },
     {
         "name": "share",
-        "unicode": "ea1b",
+        "unicode": "5564",
         "svg": "<svg t=\"1577169055872\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"4491\" width=\"200\" height=\"200\"><path d=\"M752 664c-28.5 0-54.8 10-75.4 26.7L469.4 540.8c1.7-9.3 2.6-19 2.6-28.8s-0.9-19.4-2.6-28.8l207.2-149.9C697.2 350 723.5 360 752 360c66.2 0 120-53.8 120-120s-53.8-120-120-120-120 53.8-120 120c0 11.6 1.6 22.7 4.7 33.3L439.9 415.8C410.7 377.1 364.3 352 312 352c-88.4 0-160 71.6-160 160s71.6 160 160 160c52.3 0 98.7-25.1 127.9-63.8l196.8 142.5c-3.1 10.6-4.7 21.8-4.7 33.3 0 66.2 53.8 120 120 120s120-53.8 120-120-53.8-120-120-120z m0-476c28.7 0 52 23.3 52 52s-23.3 52-52 52-52-23.3-52-52 23.3-52 52-52zM312 600c-48.5 0-88-39.5-88-88s39.5-88 88-88 88 39.5 88 88-39.5 88-88 88z m440 236c-28.7 0-52-23.3-52-52s23.3-52 52-52 52 23.3 52 52-23.3 52-52 52z\" p-id=\"4492\"></path></svg>"
     },
     {
         "name": "switch-menu",
-        "unicode": "ea1c",
+        "unicode": "5565",
         "svg": "<svg t=\"1628253142116\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"984\" width=\"200\" height=\"200\"><path d=\"M512 835.38c-78.7 0-154.49-28.67-213.41-80.74-58.39-51.6-96.11-122.37-106.21-199.26-1.79-13.6 7.79-26.07 21.38-27.85 13.59-1.79 26.07 7.79 27.85 21.38 8.54 65.01 40.45 124.86 89.86 168.52 49.83 44.04 113.94 68.29 180.52 68.29 66.58 0 130.69-24.25 180.52-68.29 49.41-43.66 81.32-103.51 89.86-168.52 1.79-13.6 14.26-23.17 27.85-21.38 13.6 1.79 23.17 14.26 21.38 27.85-10.1 76.9-47.82 147.66-106.21 199.26-58.9 52.07-134.69 80.74-213.39 80.74zM217.3 496.69c-1.13 0-2.28-0.08-3.43-0.24-13.58-1.88-23.07-14.41-21.19-28 10.57-76.4 48.46-146.64 106.7-197.78 58.79-51.62 134.3-80.05 212.62-80.05s153.83 28.43 212.62 80.05c58.24 51.14 96.13 121.38 106.7 197.78 1.88 13.58-7.61 26.12-21.19 28-13.59 1.88-26.12-7.61-28-21.19-8.94-64.59-41-124-90.28-167.27-49.73-43.66-113.6-67.71-179.86-67.71s-130.13 24.05-179.86 67.71c-49.28 43.27-81.34 102.68-90.28 167.27-1.7 12.43-12.35 21.43-24.55 21.43z\" fill=\"#333333\" p-id=\"985\"></path><path d=\"M278.98 607.83c-3.79 0-7.63-0.87-11.23-2.7l-60.96-31c-12.22-6.22-17.09-21.16-10.88-33.39 6.22-12.22 21.16-17.09 33.39-10.88l60.96 31c12.22 6.22 17.09 21.16 10.88 33.39-4.39 8.62-13.11 13.58-22.16 13.58zM805.68 496.83c-3.02 0-6.09-0.55-9.07-1.72l-63.69-25c-12.76-5.01-19.05-19.42-14.04-32.18 5.01-12.76 19.42-19.05 32.18-14.04l63.69 25c12.76 5.01 19.05 19.42 14.04 32.18-3.83 9.79-13.2 15.76-23.11 15.76z\" fill=\"#333333\" p-id=\"986\"></path><path d=\"M373.5 392m-24.83 0a24.83 24.83 0 1 0 49.66 0 24.83 24.83 0 1 0-49.66 0Z\" fill=\"#333333\" p-id=\"987\"></path><path d=\"M654 416.83H452c-13.71 0-24.83-11.12-24.83-24.83s11.12-24.83 24.83-24.83h202c13.71 0 24.83 11.12 24.83 24.83s-11.12 24.83-24.83 24.83z\" fill=\"#333333\" p-id=\"988\"></path><path d=\"M373.5 520.5m-24.83 0a24.83 24.83 0 1 0 49.66 0 24.83 24.83 0 1 0-49.66 0Z\" fill=\"#448AFF\" p-id=\"989\"></path><path d=\"M654 545.33H452c-13.71 0-24.83-11.12-24.83-24.83 0-13.71 11.12-24.83 24.83-24.83h202c13.71 0 24.83 11.12 24.83 24.83 0 13.71-11.12 24.83-24.83 24.83z\" fill=\"#448AFF\" p-id=\"990\"></path><path d=\"M373.5 649m-24.83 0a24.83 24.83 0 1 0 49.66 0 24.83 24.83 0 1 0-49.66 0Z\" fill=\"#333333\" p-id=\"991\"></path><path d=\"M654 673.83H452c-13.71 0-24.83-11.12-24.83-24.83s11.12-24.83 24.83-24.83h202c13.71 0 24.83 11.12 24.83 24.83s-11.12 24.83-24.83 24.83z\" fill=\"#333333\" p-id=\"992\"></path></svg>"
     },
     {
         "name": "sync",
-        "unicode": "ea1d",
+        "unicode": "5566",
         "svg": "<svg t=\"1577169075376\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"4608\" width=\"200\" height=\"200\"><path d=\"M168 504.2c1-43.7 10-86.1 26.9-126 17.3-41 42.1-77.7 73.7-109.4S337 212.3 378 195c42.4-17.9 87.4-27 133.9-27s91.5 9.1 133.8 27c40.9 17.3 77.7 42.1 109.3 73.8 9.9 9.9 19.2 20.4 27.8 31.4l-60.2 47c-5.3 4.1-3.5 12.5 3 14.1l175.7 43c5 1.2 9.9-2.6 9.9-7.7l0.8-180.9c0-6.7-7.7-10.5-12.9-6.3l-56.4 44.1C765.8 155.1 646.2 92 511.8 92 282.7 92 96.3 275.6 92 503.8c-0.1 4.5 3.5 8.2 8 8.2h60c4.4 0 7.9-3.5 8-7.8zM924 512h-60c-4.4 0-7.9 3.5-8 7.8-1 43.7-10 86.1-26.9 126-17.3 41-42.1 77.8-73.7 109.4S687 811.7 646 829c-42.4 17.9-87.4 27-133.9 27s-91.5-9.1-133.9-27c-40.9-17.3-77.7-42.1-109.3-73.8-9.9-9.9-19.2-20.4-27.8-31.4l60.2-47c5.3-4.1 3.5-12.5-3-14.1l-175.7-43c-5-1.2-9.9 2.6-9.9 7.7l-0.7 181c0 6.7 7.7 10.5 12.9 6.3l56.4-44.1C258.2 868.9 377.8 932 512.2 932c229.2 0 415.5-183.7 419.8-411.8 0.1-4.5-3.5-8.2-8-8.2z\" p-id=\"4609\"></path></svg>"
     },
     {
         "name": "wave",
-        "unicode": "ea1e",
+        "unicode": "5567",
         "svg": "<svg t=\"1577168879633\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2257\" width=\"200\" height=\"200\"><path d=\"M853.56074463 666.76220703a34.29052734 34.29052734 0 0 1-34.13232422-34.40917969V391.64697266c0-18.984375 15.26660156-34.40917969 34.13232422-34.40917969s34.171875 15.42480469 34.171875 34.40917969v240.70605468c0 9.13623047-3.59912109 17.87695313-10.00634766 24.32373047a34.01367188 34.01367188 0 0 1-24.12597656 10.08544922z m-170.78027344 137.55761719a34.01367188 34.01367188 0 0 1-24.12597656-10.04589844 34.52783203 34.52783203 0 0 1-10.00634766-24.32373047V254.04980469c0-18.984375 15.26660156-34.40917969 34.13232422-34.40917969s34.171875 15.42480469 34.171875 34.40917969v515.90039062c0 9.09667969-3.59912109 17.87695313-10.00634765 24.32373047a34.01367188 34.01367188 0 0 1-24.12597657 10.04589844zM512.00019776 907.5078125a34.29052734 34.29052734 0 0 1-34.171875-34.40917969V150.90136719c0-18.984375 15.30615234-34.40917969 34.171875-34.40917969s34.171875 15.42480469 34.171875 34.40917969v722.19726562c0 18.984375-15.30615234 34.40917969-34.171875 34.40917969z m-170.78027344-103.18798828a34.01367188 34.01367188 0 0 1-24.12597655-10.04589844 34.52783203 34.52783203 0 0 1-10.04589845-24.32373047V254.04980469c0-18.984375 15.30615234-34.40917969 34.171875-34.40917969s34.13232422 15.42480469 34.13232422 34.40917969v515.90039062c0 9.09667969-3.55957031 17.87695313-9.96679688 24.32373047a34.01367188 34.01367188 0 0 1-24.16552734 10.04589844zM170.43965088 666.76220703a34.29052734 34.29052734 0 0 1-34.171875-34.40917969V391.64697266c0-18.984375 15.30615234-34.40917969 34.171875-34.40917969s34.13232422 15.42480469 34.13232422 34.40917969v240.70605468c0 9.13623047-3.55957031 17.87695313-10.00634765 24.32373047a34.01367188 34.01367188 0 0 1-24.12597657 10.08544922z\" p-id=\"2258\"></path></svg>"
     },
     {
         "name": "wendangxiugai",
-        "unicode": "ea1f",
+        "unicode": "5568",
         "svg": "<svg t=\"1577168935222\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2838\" width=\"200\" height=\"200\"><path d=\"M826 472c-17.673 0-32-14.327-32-32V136a8 8 0 0 0-8-8H136a8 8 0 0 0-8 8v752a8 8 0 0 0 8 8h289c17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32H96c-17.673 0-32-14.327-32-32V96c0-17.673 14.327-32 32-32h730c17.673 0 32 14.327 32 32v344c0 17.673-14.327 32-32 32zM211 270h471c17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32H211c-17.673 0-32-14.327-32-32 0-17.673 14.327-32 32-32z m5 206h212c17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32H216c-17.673 0-32-14.327-32-32 0-17.673 14.327-32 32-32z m304 0h44c17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32h-44c-17.673 0-32-14.327-32-32 0-17.673 14.327-32 32-32z m28.821 332.061l285.434-285.434c12.497-12.496 32.758-12.496 45.255 0l70.003 70.004c12.497 12.497 12.497 32.758 0 45.255L664.08 923.32a32 32 0 0 1-21.043 9.333l-91.301 4.525c-8.826 0.437-16.335-6.363-16.772-15.188a16 16 0 0 1 0-1.584l4.525-91.302a32 32 0 0 1 9.333-21.043z m30.875 48.32l-1.4 28.902a8 8 0 0 0 8.325 8.38l29.138-1.219 267.993-267.993a8 8 0 0 0 0-11.314l-24.748-24.749a8 8 0 0 0-11.314 0L579.696 856.382zM216 746c-17.673 0-32-14.327-32-32 0-17.673 14.327-32 32-32 17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32z m104 0c-17.673 0-32-14.327-32-32 0-17.673 14.327-32 32-32 17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32z m104 0c-17.673 0-32-14.327-32-32 0-17.673 14.327-32 32-32 17.673 0 32 14.327 32 32 0 17.673-14.327 32-32 32z\" p-id=\"2839\"></path></svg>"
     },
     {
         "name": "winfo-icon-chakanbaogao",
-        "unicode": "ea20",
+        "unicode": "5569",
         "svg": "<svg t=\"1577168950499\" class=\"icon\" viewBox=\"0 0 1024 1024\" version=\"1.1\" xmlns=\"http://www.w3.org/2000/svg\" p-id=\"2955\" width=\"200\" height=\"200\"><path d=\"M687.1 303.7c0-19-15.4-34.4-34.4-34.4H307.5c-19 0-34.4 15.4-34.4 34.4s15.4 34.4 34.4 34.4h345.2c19 0 34.4-15.4 34.4-34.4zM548.8 510.7c0-19-15.4-34.4-34.4-34.4h-207c-19 0-34.4 15.4-34.4 34.4s15.4 34.4 34.4 34.4h207c19-0.1 34.4-15.4 34.4-34.4zM307.5 682.9c-19 0-34.4 15.4-34.4 34.4s15.4 34.4 34.4 34.4h69.1c19 0 34.4-15.4 34.4-34.4s-15.4-34.4-34.4-34.4h-69.1z\" p-id=\"2956\"></path><path d=\"M485.7 890.6H170.1V134.8h619.5v361.9c24.9 8.5 48.1 20.5 69 35.6v-432c0-19-15.5-34.5-34.5-34.5H135.6c-19 0-34.5 15.5-34.5 34.5v824.8c0 19 15.5 34.5 34.5 34.5h416.6c-26.2-18.8-48.8-42.2-66.5-69z\" p-id=\"2957\"></path><path d=\"M917.9 900.9L869.1 852c-0.7-0.7-1.5-1.4-2.2-2 20.4-29.9 32.3-66.1 32.3-105 0-103.2-83.6-186.8-186.8-186.8S525.5 641.8 525.5 745s83.6 186.8 186.8 186.8c39.5 0 76.1-12.3 106.3-33.2 0.6 0.7 1.2 1.4 1.9 2l48.9 48.9c13.4 13.4 35.2 13.4 48.6 0 13.4-13.4 13.4-35.2-0.1-48.6z m-122.3-72.6c-22.3 22.3-51.9 34.5-83.3 34.5-31.5 0-61.1-12.3-83.3-34.5-22.3-22.3-34.5-51.9-34.5-83.3s12.3-61.1 34.5-83.3c22.3-22.3 51.9-34.5 83.3-34.5 31.5 0 61.1 12.3 83.3 34.5 22.3 22.3 34.5 51.9 34.5 83.3s-12.2 61-34.5 83.3z\" p-id=\"2958\"></path></svg>"
     }
 ]
