@@ -1,5 +1,5 @@
 export default {
-  versions: (function() {
+  versions: (function () {
     var u = navigator.userAgent,
       app = navigator.appVersion;
     return {
@@ -13,8 +13,8 @@ export default {
       android: u.indexOf("Android") > -1 || u.indexOf("Linux") > -1, //android终端或uc浏览器
       iPhone: u.indexOf("iPhone") > -1, //是否为iPhone或者QQHD浏览器
       iPad: u.indexOf("iPad") > -1, //是否iPad
-      webApp: u.indexOf("Safari") === -1 //是否web应该程序，没有头部与底部
+      webApp: u.indexOf("Safari") === -1, //是否web应该程序，没有头部与底部
     };
   })(),
-  language: (navigator.browserLanguage || navigator.language).toLowerCase()
+  language: (navigator.browserLanguage || navigator.language).toLowerCase(),
 };
