@@ -1,5 +1,5 @@
 export default {
   SubscriptionType: "subscription",
   ServerType: "server",
-  SubscriptionServerType: "subscriptionServer"
+  SubscriptionServerType: "subscriptionServer",
 };
