@@ -5,7 +5,7 @@ import { use, registerComponent } from "buefy/src/utils/plugins";
 const Plugin = {
   install(Vue) {
     registerComponent(Vue, Input);
-  }
+  },
 };
 
 use(Plugin);
