@@ -1,6 +1,6 @@
 <template>
   <!-- TODO: mobile device compatibility -->
-  <div class="modal-card" style="width:65rem">
+  <div class="modal-card" style="width: 65rem">
     <header class="modal-card-head">
       <p class="modal-card-title">
         {{ $tc("log.logModalTitle") }}
@@ -20,14 +20,21 @@
       </b-field>
       <b-field label="Logs" style="margin-bottom: 2rem">
         <RecycleScroller
+          ref="logScroller"
           v-slot="{ item }"
           class="log-scroller"
           :items="items"
-          :item-size="28"
+          :item-size="itemSize"
+          :grid-items="1"
           :buffer="1000"
         >
           <hightlight-log class="text" :text="item.text"></hightlight-log>
         </RecycleScroller>
+      </b-field>
+      <b-field>
+        <b-switch v-model="autoScoll" @input="changeScoll">
+          {{ $tc("log.autoScoll") }}
+        </b-switch>
       </b-field>
     </section>
   </div>
@@ -43,19 +50,23 @@ export default {
       currentSkip: 0,
       intervalId: 0,
       intervalTime: 5,
-      intervalCandidate: [2, 5, 10, 15]
+      intervalCandidate: [2, 5, 10, 15],
+      itemSize: 28,
+      autoScoll: true,
     };
   },
   created() {
+    this.autoScoll = !(localStorage.getItem("log.autoScoll") === "false");
+
     this.$axios({
-      url: apiRoot + "/logger"
+      url: apiRoot + "/logger",
     }).then(this.updateLog);
   },
   mounted() {
     this.intervalId = setInterval(() => {
       this.$axios({
         url: apiRoot + `/logger`,
-        params: { skip: this.currentSkip }
+        params: { skip: this.currentSkip },
       }).then(this.updateLog);
     }, this.intervalTime * 1000);
   },
@@ -77,6 +88,9 @@ export default {
         }
         this.endOfLine = items[items.length - 1].text.endsWith("\n");
         this.currentSkip += new Blob([logs.data]).size;
+        if (this.autoScoll) {
+          this.$refs.logScroller.scrollToItem(this.items.length - 1);
+        }
       }
     },
     changeInterval(val) {
@@ -85,11 +99,14 @@ export default {
       this.intervalId = setInterval(() => {
         this.$axios({
           url: apiRoot + `/logger`,
-          params: { skip: this.currentSkip }
+          params: { skip: this.currentSkip },
         }).then(this.updateLog);
       }, this.intervalTime * 1000);
-    }
-  }
+    },
+    changeScoll(val) {
+      localStorage.setItem("log.autoScoll", val);
+    },
+  },
 };
 </script>
 <style scoped>
@@ -102,6 +119,7 @@ export default {
 <style lang="scss">
 .log-scroller {
   height: 50vh;
+
   .vue-recycle-scroller__item-wrapper {
     overflow-x: auto;
   }
