@@ -1,86 +1,38 @@
 <template>
-  <div class="modal-card" style="max-width: 520px;margin:auto">
+  <div class="modal-card" style="max-width: 520px; margin: auto">
     <header class="modal-card-head">
       <p class="modal-card-title">
         {{ $tc("configureServer.title", readonly ? 2 : 1) }}
       </p>
     </header>
     <section ref="section" :class="{ 'modal-card-body': true }">
-      <b-tabs
-        v-model="tabChoice"
-        position="is-centered"
-        class="block"
-        type="is-boxed is-twitter same-width-5"
-      >
+      <b-tabs v-model="tabChoice" position="is-centered" class="block" type="is-boxed is-twitter same-width-5">
         <b-tab-item label="V2RAY">
           <b-field label="Protocol" label-position="on-border">
-            <b-select
-              v-model="v2ray.protocol"
-              expanded
-              @input="handleV2rayProtocolSwitch"
-            >
+            <b-select v-model="v2ray.protocol" expanded @input="handleV2rayProtocolSwitch">
               <option value="vmess">VMESS</option>
               <option value="vless">VLESS</option>
             </b-select>
           </b-field>
           <b-field label="Name" label-position="on-border">
-            <b-input
-              ref="v2ray_name"
-              v-model="v2ray.ps"
-              :placeholder="$t('configureServer.servername')"
-              expanded
-            />
+            <b-input ref="v2ray_name" v-model="v2ray.ps" :placeholder="$t('configureServer.servername')" expanded />
           </b-field>
           <b-field label="Host" label-position="on-border">
-            <b-input
-              ref="v2ray_add"
-              v-model="v2ray.add"
-              required
-              placeholder="IP / HOST"
-              expanded
-            />
+            <b-input ref="v2ray_add" v-model="v2ray.add" required placeholder="IP / HOST" expanded />
           </b-field>
           <b-field label="Port" label-position="on-border">
-            <b-input
-              ref="v2ray_port"
-              v-model="v2ray.port"
-              required
-              :placeholder="$t('configureServer.port')"
-              type="number"
-              expanded
-            />
+            <b-input ref="v2ray_port" v-model="v2ray.port" required :placeholder="$t('configureServer.port')"
+              type="number" expanded />
           </b-field>
           <b-field label="ID" label-position="on-border">
-            <b-input
-              ref="v2ray_id"
-              v-model="v2ray.id"
-              required
-              placeholder="UserID"
-              expanded
-            />
+            <b-input ref="v2ray_id" v-model="v2ray.id" required placeholder="UserID" expanded />
           </b-field>
-          <b-field
-            v-if="v2ray.protocol === 'vmess'"
-            label="AlterID"
-            label-position="on-border"
-          >
-            <b-input
-              ref="v2ray_aid"
-              v-model="v2ray.aid"
-              placeholder="AlterID"
-              type="number"
-              min="0"
-              max="65535"
-              required
-              expanded
-            />
+          <b-field v-if="v2ray.protocol === 'vmess'" label="AlterID" label-position="on-border">
+            <b-input ref="v2ray_aid" v-model="v2ray.aid" placeholder="AlterID" type="number" min="0" max="65535"
+              expanded />
           </b-field>
-          <b-field
-            v-if="v2ray.protocol === 'vmess'"
-            label="Security"
-            label-position="on-border"
-          >
-            <b-select v-model="v2ray.scy" expanded>
+          <b-field v-if="v2ray.protocol === 'vmess'" label="Security" label-position="on-border">
+            <b-select v-model="v2ray.scy" expanded required>
               <option value="auto">Auto</option>
               <option value="aes-128-gcm">aes-128-gcm</option>
               <option value="chacha20-poly1305">chacha20-poly1305</option>
@@ -88,51 +40,54 @@
               <option value="zero">zero</option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="v2ray.type !== 'dtls'"
-            label="TLS"
-            label-position="on-border"
-          >
+          <b-field v-show="v2ray.type !== 'dtls'" label="TLS" label-position="on-border">
             <b-select v-model="v2ray.tls" expanded @input="handleNetworkChange">
               <option value="none">{{ $t("setting.options.off") }}</option>
               <option value="tls">tls</option>
+              <option v-if="variant() === 'xray'" value="reality">reality</option>
+              <option v-if="variant() === 'xray'" value="xtls">xtls</option>
             </b-select>
+          </b-field>
+          <b-field v-if="v2ray.tls !== 'none'" label="SNI" label-position="on-border">
+            <b-input ref="v2ray_sni" v-model="v2ray.sni" placeholder="SNI" expanded />
+          </b-field>
+          <b-field v-show="v2ray.tls === 'tls' || v2ray.tls === 'reality'" label="uTLS fingerprint"
+            label-position="on-border">
+            <b-input ref="v2ray_fp" v-model="v2ray.fp" placeholder="A uTLS compatable fingerprint name" expanded />
+          </b-field>
+          <b-field v-if="v2ray.protocol === 'vless' && v2ray.tls !== 'none'" ref="v2ray_flow" label="Flow"
+            label-position="on-border">
+            <b-input ref="v2ray_flow" v-model="v2ray.flow" placeholder="Flow" expanded />
+          </b-field>
+          <b-field v-show="v2ray.tls === 'reality'" label="pbk" label-position="on-border">
+            <b-input v-model="v2ray.pbk" placeholder="pbk" expanded />
+          </b-field>
+          <b-field v-show="v2ray.tls === 'reality'" label="sid" label-position="on-border">
+            <b-input v-model="v2ray.sid" placeholder="sid" expanded />
+          </b-field>
+          <b-field v-show="v2ray.tls === 'reality'" label="spx" label-position="on-border">
+            <b-input v-model="v2ray.spx" placeholder="spx" expanded />
           </b-field>
           <b-field v-show="v2ray.tls !== 'none'" label-position="on-border">
             <template slot="label">
               AllowInsecure
-              <b-tooltip
-                v-show="v2ray.protocol === 'vless'"
-                type="is-dark"
-                :label="$t('server.messages.notRecommend', { name: 'VLESS' })"
-                multilined
-                position="is-right"
-              >
-                <b-icon
-                  size="is-small"
-                  icon=" iconfont icon-help-circle-outline"
-                  style="position:relative;top:2px;right:3px;font-weight:normal"
-                />
+              <b-tooltip v-show="v2ray.protocol === 'vless'" type="is-dark"
+                :label="$t('server.messages.notRecommend', { name: 'VLESS' })" multilined position="is-right">
+                <b-icon size="is-small" icon=" iconfont icon-help-circle-outline" style="
+                    position: relative;
+                    top: 2px;
+                    right: 3px;
+                    font-weight: normal;
+                  " />
               </b-tooltip>
             </template>
-            <b-select
-              ref="v2ray_allow_insecure"
-              v-model="v2ray.allowInsecure"
-              expanded
-              required
-            >
+            <b-select ref="v2ray_allow_insecure" v-model="v2ray.allowInsecure" expanded required>
               <option :value="false">{{ $t("operations.no") }}</option>
               <option :value="true">{{ $t("operations.yes") }}</option>
             </b-select>
           </b-field>
           <b-field label="Network" label-position="on-border">
-            <b-select
-              ref="v2ray_net"
-              v-model="v2ray.net"
-              expanded
-              required
-              @input="handleNetworkChange"
-            >
+            <b-select ref="v2ray_net" v-model="v2ray.net" expanded required @input="handleNetworkChange">
               <option value="tcp">TCP</option>
               <option value="kcp">mKCP</option>
               <option value="ws">WebSocket</option>
@@ -140,157 +95,88 @@
               <option value="grpc">gRPC</option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="v2ray.net === 'tcp'"
-            label="Type"
-            label-position="on-border"
-          >
+          <b-field v-show="v2ray.net === 'tcp'" label="Type" label-position="on-border">
             <b-select v-model="v2ray.type" expanded>
-              <option value="none"
-                >{{ $t("configureServer.noObfuscation") }}
+              <option value="none">
+                {{ $t("configureServer.noObfuscation") }}
               </option>
-              <option value="http"
-                >{{ $t("configureServer.httpObfuscation") }}
+              <option value="http">
+                {{ $t("configureServer.httpObfuscation") }}
               </option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="v2ray.net === 'kcp'"
-            label="Type"
-            label-position="on-border"
-          >
+          <b-field v-show="v2ray.net === 'kcp'" label="Type" label-position="on-border">
             <b-select v-model="v2ray.type" expanded>
-              <option value="none"
-                >{{ $t("configureServer.noObfuscation") }}
+              <option value="none">
+                {{ $t("configureServer.noObfuscation") }}
               </option>
-              <option value="srtp"
-                >{{ $t("configureServer.srtpObfuscation") }}
+              <option value="srtp">
+                {{ $t("configureServer.srtpObfuscation") }}
               </option>
-              <option value="utp"
-                >{{ $t("configureServer.utpObfuscation") }}
+              <option value="utp">
+                {{ $t("configureServer.utpObfuscation") }}
               </option>
-              <option value="wechat-video"
-                >{{ $t("configureServer.wechatVideoObfuscation") }}
+              <option value="wechat-video">
+                {{ $t("configureServer.wechatVideoObfuscation") }}
               </option>
-              <option value="dtls"
-                >{{
+              <option value="dtls">
+                {{
                   `${$t("configureServer.dtlsObfuscation")}(${$t(
                     "configureServer.forceTLS"
                   )})`
                 }}
               </option>
-              <option value="wireguard"
-                >{{ $t("configureServer.wireguardObfuscation") }}
+              <option value="wireguard">
+                {{ $t("configureServer.wireguardObfuscation") }}
               </option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="
-              v2ray.net === 'ws' ||
-                v2ray.net === 'h2' ||
-                v2ray.tls === 'tls' ||
-                (v2ray.net === 'tcp' && v2ray.type === 'http')
-            "
-            label="Host"
-            label-position="on-border"
-          >
-            <b-input
-              v-model="v2ray.host"
-              :placeholder="$t('configureServer.hostObfuscation')"
-              expanded
-            />
+          <b-field v-show="v2ray.net === 'ws' ||
+            v2ray.net === 'h2' ||
+            v2ray.tls === 'tls' ||
+            (v2ray.net === 'tcp' && v2ray.type === 'http')
+            " label="Host" label-position="on-border">
+            <b-input v-model="v2ray.host" :placeholder="$t('configureServer.hostObfuscation')" expanded />
           </b-field>
-          <b-field
-            v-show="v2ray.tls === 'tls'"
-            label="Alpn"
-            label-position="on-border"
-          >
+          <b-field v-show="v2ray.tls === 'tls'" label="Alpn" label-position="on-border">
             <b-input v-model="v2ray.alpn" placeholder="h2,http/1.1" expanded />
           </b-field>
-          <b-field
-            v-show="
-              v2ray.net === 'ws' ||
-                v2ray.net === 'h2' ||
-                (v2ray.net === 'tcp' && v2ray.type === 'http')
-            "
-            label="Path"
-            label-position="on-border"
-          >
-            <b-input
-              v-model="v2ray.path"
-              :placeholder="$t('configureServer.pathObfuscation')"
-              expanded
-            />
+          <b-field v-show="v2ray.net === 'ws' ||
+            v2ray.net === 'h2' ||
+            (v2ray.net === 'tcp' && v2ray.type === 'http')
+            " label="Path" label-position="on-border">
+            <b-input v-model="v2ray.path" :placeholder="$t('configureServer.pathObfuscation')" expanded />
           </b-field>
-          <b-field
-            v-show="v2ray.net === 'mkcp' || v2ray.net === 'kcp'"
-            label="Seed"
-            label-position="on-border"
-          >
-            <b-input
-              v-model="v2ray.path"
-              :placeholder="$t('configureServer.seedObfuscation')"
-              expanded
-            />
+          <b-field v-show="v2ray.net === 'mkcp' || v2ray.net === 'kcp'" label="Seed" label-position="on-border">
+            <b-input v-model="v2ray.path" :placeholder="$t('configureServer.seedObfuscation')" expanded />
           </b-field>
-          <b-field
-            v-show="v2ray.net === 'grpc'"
-            label="ServiceName"
-            label-position="on-border"
-          >
-            <b-input
-              ref="v2ray_service_name"
-              v-model="v2ray.path"
-              type="text"
-              expanded
-            />
+          <b-field v-show="v2ray.net === 'grpc'" label="ServiceName" label-position="on-border">
+            <b-input ref="v2ray_service_name" v-model="v2ray.path" type="text" expanded />
           </b-field>
         </b-tab-item>
         <b-tab-item label="SS">
           <b-field label="Name" label-position="on-border">
-            <b-input
-              ref="ss_name"
-              v-model="ss.name"
-              :placeholder="$t('configureServer.servername')"
-              expanded
-            />
+            <b-input ref="ss_name" v-model="ss.name" :placeholder="$t('configureServer.servername')" expanded />
           </b-field>
           <b-field label="Host" label-position="on-border">
-            <b-input
-              ref="ss_server"
-              v-model="ss.server"
-              required
-              placeholder="IP / HOST"
-              expanded
-            />
+            <b-input ref="ss_server" v-model="ss.server" required placeholder="IP / HOST" expanded />
           </b-field>
           <b-field label="Port" label-position="on-border">
-            <b-input
-              ref="ss_port"
-              v-model="ss.port"
-              required
-              :placeholder="$t('configureServer.port')"
-              type="number"
-              expanded
-            />
+            <b-input ref="ss_port" v-model="ss.port" required :placeholder="$t('configureServer.port')" type="number"
+              expanded />
           </b-field>
           <b-field label="Password" label-position="on-border">
-            <b-input
-              ref="ss_password"
-              v-model="ss.password"
-              required
-              :placeholder="$t('configureServer.password')"
-              expanded
-            />
+            <b-input ref="ss_password" v-model="ss.password" required :placeholder="$t('configureServer.password')"
+              expanded />
           </b-field>
           <b-field label="Method" label-position="on-border">
             <b-select ref="ss_method" v-model="ss.method" expanded required>
               <option value="aes-128-gcm">aes-128-gcm</option>
               <option value="aes-256-gcm">aes-256-gcm</option>
               <option value="chacha20-poly1305">chacha20-poly1305</option>
-              <option value="chacha20-ietf-poly1305"
-                >chacha20-ietf-poly1305</option
-              >
+              <option value="chacha20-ietf-poly1305">
+                chacha20-ietf-poly1305
+              </option>
               <option value="plain">plain</option>
               <option value="none">none</option>
             </b-select>
@@ -302,24 +188,17 @@
               <option value="v2ray-plugin">v2ray-plugin</option>
             </b-select>
           </b-field>
-          <b-field
-            v-if="ss.plugin === 'simple-obfs' || ss.plugin === 'v2ray-plugin'"
-            label-position="on-border"
-            class="with-icon-alert"
-          >
+          <b-field v-if="ss.plugin === 'simple-obfs' || ss.plugin === 'v2ray-plugin'" label-position="on-border"
+            class="with-icon-alert">
             <template slot="label">
               Impl
-              <b-tooltip
-                type="is-dark"
-                :label="$t('setting.messages.ssPluginImpl')"
-                multilined
-                position="is-right"
-              >
-                <b-icon
-                  size="is-samll"
-                  icon=" iconfont icon-help-circle-outline"
-                  style="position:relative;top:2px;right:3px;font-weight:normal"
-                />
+              <b-tooltip type="is-dark" :label="$t('setting.messages.ssPluginImpl')" multilined position="is-right">
+                <b-icon size="is-samll" icon=" iconfont icon-help-circle-outline" style="
+                    position: relative;
+                    top: 2px;
+                    right: 3px;
+                    font-weight: normal;
+                  " />
               </b-tooltip>
             </template>
             <b-select ref="ss_plugin_impl" v-model="ss.impl" expanded>
@@ -328,98 +207,49 @@
               <option value="transport">transport</option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="ss.plugin === 'simple-obfs'"
-            label="Obfs"
-            label-position="on-border"
-          >
+          <b-field v-show="ss.plugin === 'simple-obfs'" label="Obfs" label-position="on-border">
             <b-select ref="ss_obfs" v-model="ss.obfs" expanded>
               <option value="http">http</option>
               <option value="tls">tls</option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="ss.plugin === 'v2ray-plugin'"
-            label="Mode"
-            label-position="on-border"
-          >
+          <b-field v-show="ss.plugin === 'v2ray-plugin'" label="Mode" label-position="on-border">
             <b-select ref="ss_mode" v-model="ss.mode" expanded>
               <option value="websocket">websocket</option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="ss.plugin === 'v2ray-plugin'"
-            label="TLS"
-            label-position="on-border"
-          >
+          <b-field v-show="ss.plugin === 'v2ray-plugin'" label="TLS" label-position="on-border">
             <b-select ref="ss_tls" v-model="ss.tls" expanded>
               <option value="">{{ $t("setting.options.off") }}</option>
               <option value="tls">tls</option>
             </b-select>
           </b-field>
-          <b-field
-            v-if="
-              (ss.plugin === 'simple-obfs' &&
-                (ss.obfs === 'http' || ss.obfs === 'tls')) ||
-                ss.plugin === 'v2ray-plugin'
-            "
-            label="Host"
-            label-position="on-border"
-          >
-            <b-input
-              ref="ss_host"
-              v-model="ss.host"
-              placeholder="(optional)"
-              expanded
-            />
+          <b-field v-if="(ss.plugin === 'simple-obfs' &&
+            (ss.obfs === 'http' || ss.obfs === 'tls')) ||
+            ss.plugin === 'v2ray-plugin'
+            " label="Host" label-position="on-border">
+            <b-input ref="ss_host" v-model="ss.host" placeholder="(optional)" expanded />
           </b-field>
-          <b-field
-            v-if="
-              (ss.plugin === 'simple-obfs' && ss.obfs === 'http') ||
-                ss.plugin === 'v2ray-plugin'
-            "
-            label="Path"
-            label-position="on-border"
-          >
+          <b-field v-if="(ss.plugin === 'simple-obfs' && ss.obfs === 'http') ||
+            ss.plugin === 'v2ray-plugin'
+            " label="Path" label-position="on-border">
             <b-input ref="ss_path" v-model="ss.path" placeholder="/" expanded />
           </b-field>
         </b-tab-item>
         <b-tab-item label="SSR">
           <b-field label="Name" label-position="on-border">
-            <b-input
-              ref="ssr_name"
-              v-model="ssr.name"
-              :placeholder="$t('configureServer.servername')"
-              expanded
-            />
+            <b-input ref="ssr_name" v-model="ssr.name" :placeholder="$t('configureServer.servername')" expanded />
           </b-field>
           <b-field label="Host" label-position="on-border">
-            <b-input
-              ref="ssr_server"
-              v-model="ssr.server"
-              required
-              placeholder="IP / HOST"
-              expanded
-            />
+            <b-input ref="ssr_server" v-model="ssr.server" required placeholder="IP / HOST" expanded />
           </b-field>
           <b-field label="Port" label-position="on-border">
-            <b-input
-              ref="ssr_port"
-              v-model="ssr.port"
-              required
-              :placeholder="$t('configureServer.port')"
-              type="number"
-              expanded
-            />
+            <b-input ref="ssr_port" v-model="ssr.port" required :placeholder="$t('configureServer.port')" type="number"
+              expanded />
           </b-field>
           <b-field label="Password" label-position="on-border">
-            <b-input
-              ref="ssr_password"
-              v-model="ssr.password"
-              required
-              :placeholder="$t('configureServer.password')"
-              expanded
-            />
+            <b-input ref="ssr_password" v-model="ssr.password" required :placeholder="$t('configureServer.password')"
+              expanded />
           </b-field>
           <b-field label="Method" label-position="on-border">
             <b-select ref="ssr_method" v-model="ssr.method" expanded required>
@@ -459,17 +289,8 @@
               <option value="auth_chain_b">auth_chain_b</option>
             </b-select>
           </b-field>
-          <b-field
-            v-if="ssr.proto !== 'origin'"
-            label="Protocol Param"
-            label-position="on-border"
-          >
-            <b-input
-              ref="ssr_protoParam"
-              v-model="ssr.protoParam"
-              placeholder="(optional)"
-              expanded
-            />
+          <b-field v-if="ssr.proto !== 'origin'" label="Protocol Param" label-position="on-border">
+            <b-input ref="ssr_protoParam" v-model="ssr.protoParam" placeholder="(optional)" expanded />
           </b-field>
           <b-field label="Obfs" label-position="on-border">
             <b-select ref="ssr_obfs" v-model="ssr.obfs" expanded required>
@@ -480,130 +301,63 @@
               <option value="tls1.2_ticket_auth">tls1.2_ticket_auth</option>
             </b-select>
           </b-field>
-          <b-field
-            v-if="ssr.obfs !== 'plain'"
-            label="Obfs Param"
-            label-position="on-border"
-          >
-            <b-input
-              ref="ssr_obfsParam"
-              v-model="ssr.obfsParam"
-              placeholder="(optional)"
-              expanded
-            />
+          <b-field v-if="ssr.obfs !== 'plain'" label="Obfs Param" label-position="on-border">
+            <b-input ref="ssr_obfsParam" v-model="ssr.obfsParam" placeholder="(optional)" expanded />
           </b-field>
         </b-tab-item>
         <b-tab-item label="Trojan">
           <b-field label="Name" label-position="on-border">
-            <b-input
-              ref="trojan_name"
-              v-model="trojan.name"
-              :placeholder="$t('configureServer.servername')"
-              expanded
-            />
+            <b-input ref="trojan_name" v-model="trojan.name" :placeholder="$t('configureServer.servername')" expanded />
           </b-field>
           <b-field label="Host" label-position="on-border">
-            <b-input
-              ref="trojan_server"
-              v-model="trojan.server"
-              required
-              placeholder="IP / HOST"
-              expanded
-            />
+            <b-input ref="trojan_server" v-model="trojan.server" required placeholder="IP / HOST" expanded />
           </b-field>
           <b-field label="Port" label-position="on-border">
-            <b-input
-              ref="trojan_port"
-              v-model="trojan.port"
-              required
-              :placeholder="$t('configureServer.port')"
-              type="number"
-              expanded
-            />
+            <b-input ref="trojan_port" v-model="trojan.port" required :placeholder="$t('configureServer.port')"
+              type="number" expanded />
           </b-field>
           <b-field label="Password" label-position="on-border">
-            <b-input
-              ref="trojan_password"
-              v-model="trojan.password"
-              required
-              :placeholder="$t('configureServer.password')"
-              expanded
-            />
+            <b-input ref="trojan_password" v-model="trojan.password" required
+              :placeholder="$t('configureServer.password')" expanded />
           </b-field>
           <b-field label="Protocol" label-position="on-border">
-            <b-select
-              ref="trojan_method"
-              v-model="trojan.method"
-              expanded
-              required
-            >
+            <b-select ref="trojan_method" v-model="trojan.method" expanded required>
               <option value="origin">{{ $t("configureServer.origin") }}</option>
               <option value="shadowsocks">shadowsocks</option>
             </b-select>
           </b-field>
-          <b-field
-            v-if="trojan.method === 'shadowsocks'"
-            label="Shadowsocks Cipher"
-            label-position="on-border"
-          >
-            <b-select
-              ref="trojan_ss_cipher"
-              v-model="trojan.ssCipher"
-              expanded
-              required
-            >
+          <b-field v-if="trojan.method === 'shadowsocks'" label="Shadowsocks Cipher" label-position="on-border">
+            <b-select ref="trojan_ss_cipher" v-model="trojan.ssCipher" expanded required>
               <option value="aes-128-gcm">aes-128-gcm</option>
               <option value="aes-256-gcm">aes-256-gcm</option>
               <option value="chacha20-poly1305">chacha20-poly1305</option>
-              <option value="chacha20-ietf-poly1305"
-                >chacha20-ietf-poly1305</option
-              >
+              <option value="chacha20-ietf-poly1305">
+                chacha20-ietf-poly1305
+              </option>
             </b-select>
           </b-field>
-          <b-field
-            v-if="trojan.method === 'shadowsocks'"
-            label="Shadowsocks Password"
-            label-position="on-border"
-          >
-            <b-input
-              ref="trojan_ss_password"
-              v-model="trojan.ssPassword"
-              required
-              :placeholder="`shadowsocks${$t('configureServer.password')}`"
-              expanded
-            />
+          <b-field v-if="trojan.method === 'shadowsocks'" label="Shadowsocks Password" label-position="on-border">
+            <b-input ref="trojan_ss_password" v-model="trojan.ssPassword" required
+              :placeholder="`shadowsocks${$t('configureServer.password')}`" expanded />
           </b-field>
           <b-field label-position="on-border">
             <template slot="label">
               AllowInsecure
-              <b-tooltip
-                v-show="trojan.method !== 'origin' || trojan.obfs !== 'none'"
-                type="is-dark"
-                :label="
-                  $t('server.messages.notAllowInsecure', { name: 'Trojan-Go' })
-                "
-                multilined
-                position="is-right"
-              >
-                <b-icon
-                  size="is-small"
-                  icon=" iconfont icon-help-circle-outline"
-                  style="position:relative;top:2px;right:3px;font-weight:normal"
-                />
+              <b-tooltip v-show="trojan.method !== 'origin' || trojan.obfs !== 'none'" type="is-dark" :label="$t('server.messages.notAllowInsecure', { name: 'Trojan-Go' })
+                " multilined position="is-right">
+                <b-icon size="is-small" icon=" iconfont icon-help-circle-outline" style="
+                    position: relative;
+                    top: 2px;
+                    right: 3px;
+                    font-weight: normal;
+                  " />
               </b-tooltip>
             </template>
-            <b-select
-              ref="trojan_allow_insecure"
-              v-model="trojan.allowInsecure"
-              expanded
-              required
-            >
+            <b-select ref="trojan_allow_insecure" v-model="trojan.allowInsecure" expanded required>
               <option :value="false">{{ $t("operations.no") }}</option>
-              <option
-                :value="true"
-                :disabled="trojan.method !== 'origin' || trojan.obfs !== 'none'"
-                >{{ $t("operations.yes") }}</option
-              >
+              <option :value="true" :disabled="trojan.method !== 'origin' || trojan.obfs !== 'none'">
+                {{ $t("operations.yes") }}
+              </option>
             </b-select>
           </b-field>
           <b-field label="SNI(Peer)" label-position="on-border">
@@ -611,25 +365,123 @@
           </b-field>
           <b-field label="Obfs" label-position="on-border">
             <b-select ref="trojan_obfs" v-model="trojan.obfs" expanded required>
-              <option value="none">{{
-                $t("configureServer.noObfuscation")
-              }}</option>
+              <option value="none">
+                {{ $t("configureServer.noObfuscation") }}
+              </option>
               <option value="websocket">websocket</option>
             </b-select>
           </b-field>
-          <b-field
-            v-show="trojan.obfs === 'websocket'"
-            label="Websocket Host"
-            label-position="on-border"
-          >
+          <b-field v-show="trojan.obfs === 'websocket'" label="Websocket Host" label-position="on-border">
             <b-input v-model="trojan.host" expanded />
           </b-field>
-          <b-field
-            v-show="trojan.obfs === 'websocket'"
-            label="Websocket Path"
-            label-position="on-border"
-          >
+          <b-field v-show="trojan.obfs === 'websocket'" label="Websocket Path" label-position="on-border">
             <b-input v-model="trojan.path" placeholder="/" expanded />
+          </b-field>
+        </b-tab-item>
+
+        <b-tab-item label="Juicity">
+          <b-field label="Name" label-position="on-border">
+            <b-input ref="juicity_name" v-model="juicity.name" :placeholder="$t('configureServer.servername')" expanded />
+          </b-field>
+          <b-field label="Host" label-position="on-border">
+            <b-input ref="juicity_server" v-model="juicity.server" required placeholder="IP / HOST" expanded />
+          </b-field>
+          <b-field label="Port" label-position="on-border">
+            <b-input ref="juicity_port" v-model="juicity.port" required :placeholder="$t('configureServer.port')"
+              type="number" expanded />
+          </b-field>
+          <b-field label="UUID" label-position="on-border">
+            <b-input ref="juicity_uuid" v-model="juicity.uuid" required placeholder="UUID" expanded />
+          </b-field>
+          <b-field label="Password" label-position="on-border">
+            <b-input ref="juicity_password" v-model="juicity.password" required
+              :placeholder="$t('configureServer.password')" expanded />
+          </b-field>
+          <b-field label="Congestion Control" label-position="on-border">
+            <b-select ref="juicity_cc" v-model="juicity.cc" expanded required>
+              <option value="bbr">bbr</option>
+            </b-select>
+          </b-field>
+          <b-field label-position="on-border">
+            <template slot="label">
+              AllowInsecure
+            </template>
+            <b-select ref="juicity_allow_insecure" v-model="juicity.allowInsecure" expanded required>
+              <option :value="false">{{ $t("operations.no") }}</option>
+              <option :value="true">
+                {{ $t("operations.yes") }}
+              </option>
+            </b-select>
+          </b-field>
+          <b-field label="SNI" label-position="on-border">
+            <b-input v-model="juicity.sni" placeholder="SNI" expanded />
+          </b-field>
+          <b-field label="Pinned Cert Chain Sha256" label-position="on-border">
+            <b-input v-model="juicity.pinnedCertchainSha256" :placeholder="$t('configureServer.pinnedCertchainSha256')"
+              expanded />
+          </b-field>
+        </b-tab-item>
+
+        <b-tab-item label="Tuic">
+          <b-field label="Name" label-position="on-border">
+            <b-input ref="tuic_name" v-model="tuic.name" :placeholder="$t('configureServer.servername')" expanded />
+          </b-field>
+          <b-field label="Host" label-position="on-border">
+            <b-input ref="tuic_server" v-model="tuic.server" required placeholder="IP / HOST" expanded />
+          </b-field>
+          <b-field label="Port" label-position="on-border">
+            <b-input ref="tuic_port" v-model="tuic.port" required :placeholder="$t('configureServer.port')" type="number"
+              expanded />
+          </b-field>
+          <b-field label="UUID" label-position="on-border">
+            <b-input ref="tuic_uuid" v-model="tuic.uuid" required placeholder="UUID" expanded />
+          </b-field>
+          <b-field label="Password" label-position="on-border">
+            <b-input ref="tuic_password" v-model="tuic.password" required :placeholder="$t('configureServer.password')"
+              expanded />
+          </b-field>
+          <b-field label="Congestion Control" label-position="on-border">
+            <b-select ref="tuic_cc" v-model="tuic.cc" expanded required>
+              <option value="bbr">bbr</option>
+            </b-select>
+          </b-field>
+          <b-field label-position="on-border">
+            <template slot="label">
+              AllowInsecure
+            </template>
+            <b-select ref="tuic_allow_insecure" v-if="tuic.disableSni === false" v-model="tuic.allowInsecure" expanded
+              required>
+              <option :value="false">{{ $t("operations.no") }}</option>
+              <option :value="true">
+                {{ $t("operations.yes") }}
+              </option>
+            </b-select>
+          </b-field>
+          <b-field label-position="on-border">
+            <template slot="label">
+              DisableSni
+            </template>
+            <b-select ref="tuic_disable_sni" v-model="tuic.disableSni" expanded required>
+              <option :value="false">{{ $t("operations.no") }}</option>
+              <option :value="true">
+                {{ $t("operations.yes") }}
+              </option>
+            </b-select>
+          </b-field>
+          <b-field label="SNI" label-position="on-border" v-if="tuic.disableSni === false">
+            <b-input v-model="tuic.sni" placeholder="SNI" expanded />
+          </b-field>
+          <b-field label="ALPN" label-position="on-border">
+            <b-input v-model="tuic.alpn" placeholder="h3" expanded />
+          </b-field>
+          <b-field label-position="on-border">
+            <template slot="label">
+              UDP relay mode
+            </template>
+            <b-select ref="tuic_udp_relay_mode" v-model="tuic.udpRelayMode" expanded required>
+              <option value="native">native</option>
+              <option value="quic">quic</option>
+            </b-select>
           </b-field>
         </b-tab-item>
 
@@ -641,93 +493,41 @@
             </b-select>
           </b-field>
           <b-field label="Name" label-position="on-border">
-            <b-input
-              ref="http_name"
-              v-model="http.name"
-              :placeholder="$t('configureServer.servername')"
-              expanded
-            />
+            <b-input ref="http_name" v-model="http.name" :placeholder="$t('configureServer.servername')" expanded />
           </b-field>
           <b-field label="Host" label-position="on-border">
-            <b-input
-              ref="http_host"
-              v-model="http.host"
-              required
-              placeholder="IP / HOST"
-              expanded
-            />
+            <b-input ref="http_host" v-model="http.host" required placeholder="IP / HOST" expanded />
           </b-field>
           <b-field label="Port" label-position="on-border">
-            <b-input
-              ref="http_port"
-              v-model="http.port"
-              required
-              :placeholder="$t('configureServer.port')"
-              type="number"
-              expanded
-            />
+            <b-input ref="http_port" v-model="http.port" required :placeholder="$t('configureServer.port')" type="number"
+              expanded />
           </b-field>
           <b-field label="Username" label-position="on-border">
-            <b-input
-              ref="http_username"
-              v-model="http.username"
-              :placeholder="$t('configureServer.username')"
-              expanded
-            />
+            <b-input ref="http_username" v-model="http.username" :placeholder="$t('configureServer.username')" expanded />
           </b-field>
           <b-field label="Password" label-position="on-border">
-            <b-input
-              ref="http_password"
-              v-model="http.password"
-              :placeholder="$t('configureServer.password')"
-              expanded
-            />
+            <b-input ref="http_password" v-model="http.password" :placeholder="$t('configureServer.password')" expanded />
           </b-field>
         </b-tab-item>
 
         <b-tab-item label="SOCKS5">
           <b-field label="Name" label-position="on-border">
-            <b-input
-              ref="socks5_name"
-              v-model="socks5.name"
-              :placeholder="$t('configureServer.servername')"
-              expanded
-            />
+            <b-input ref="socks5_name" v-model="socks5.name" :placeholder="$t('configureServer.servername')" expanded />
           </b-field>
           <b-field label="Host" label-position="on-border">
-            <b-input
-              ref="socks5_host"
-              v-model="socks5.host"
-              required
-              placeholder="IP / HOST"
-              expanded
-            />
+            <b-input ref="socks5_host" v-model="socks5.host" required placeholder="IP / HOST" expanded />
           </b-field>
           <b-field label="Port" label-position="on-border">
-            <b-input
-              ref="socks5_port"
-              v-model="socks5.port"
-              required
-              :placeholder="$t('configureServer.port')"
-              type="number"
-              expanded
-            />
+            <b-input ref="socks5_port" v-model="socks5.port" required :placeholder="$t('configureServer.port')"
+              type="number" expanded />
           </b-field>
           <b-field label="Username" label-position="on-border">
-            <b-input
-              ref="socks5_username"
-              v-model="socks5.username"
-              :placeholder="$t('configureServer.username')"
-              expanded
-            />
+            <b-input ref="socks5_username" v-model="socks5.username" :placeholder="$t('configureServer.username')"
+              expanded />
           </b-field>
           <b-field label="Password" label-position="on-border">
-            <b-input
-              ref="socks5_password"
-              v-model="socks5.password"
-              :placeholder="$t('configureServer.password')"
-              expanded
-            />
+            <b-input ref="socks5_password" v-model="socks5.password" :placeholder="$t('configureServer.password')"
+              expanded />
           </b-field>
         </b-tab-item>
       </b-tabs>
@@ -744,9 +544,8 @@
 </template>
 
 <script>
-import { handleResponse } from "@/assets/js/utils";
+import { generateURL, handleResponse, parseURL } from "@/assets/js/utils";
 import { Base64 } from "js-base64";
-import { parseURL, generateURL } from "@/assets/js/utils";
 
 export default {
   name: "ModalServer",
@@ -755,12 +554,12 @@ export default {
       type: Object,
       default() {
         return null;
-      }
+      },
     },
     readonly: {
       type: Boolean,
-      default: false
-    }
+      default: false,
+    },
   },
   data: () => ({
     vlessVersion: 0,
@@ -769,17 +568,22 @@ export default {
       add: "",
       port: "",
       id: "",
+      flow: "",
       aid: "",
       net: "tcp",
       type: "none",
       host: "",
       path: "",
       tls: "none",
+      fp: "",
+      pbk: "",
+      sid: "",
+      spx: "",
       alpn: "",
-      scy: "",
+      scy: "auto",
       v: "",
       allowInsecure: false,
-      protocol: "vmess"
+      protocol: "vmess",
     },
     ss: {
       method: "aes-128-gcm",
@@ -794,7 +598,7 @@ export default {
       port: "",
       name: "",
       protocol: "ss",
-      impl: ""
+      impl: "",
     },
     ssr: {
       method: "aes-128-cfb",
@@ -806,7 +610,7 @@ export default {
       protoParam: "",
       obfs: "plain",
       obfsParam: "",
-      protocol: "ssr"
+      protocol: "ssr",
     },
     trojan: {
       name: "",
@@ -821,7 +625,33 @@ export default {
       ssCipher: "aes-128-gcm",
       ssPassword: "",
       obfs: "none" /* websocket */,
-      protocol: "trojan"
+      protocol: "trojan",
+    },
+    juicity: {
+      name: "",
+      server: "",
+      port: "",
+      sni: "",
+      cc: "bbr",
+      uuid: "",
+      password: "",
+      pinnedCertchainSha256: "",
+      allowInsecure: false,
+      protocol: "juicity"
+    },
+    tuic: {
+      name: "",
+      server: "",
+      port: "",
+      sni: "",
+      cc: "bbr",
+      uuid: "",
+      password: "",
+      allowInsecure: false,
+      disableSni: false,
+      alpn: "h3",
+      udpRelayMode: "native",
+      protocol: "tuic"
     },
     http: {
       username: "",
@@ -829,7 +659,7 @@ export default {
       host: "",
       port: "",
       protocol: "http",
-      name: ""
+      name: "",
     },
     socks5: {
       username: "",
@@ -837,26 +667,19 @@ export default {
       host: "",
       port: "",
       protocol: "socks5",
-      name: ""
+      name: "",
     },
-    tabChoice: 0
+    tabChoice: 0,
   }),
-  computed: {
-    filteredDataArray() {
-      return this.presetFlows.filter(option => {
-        return option.toString().indexOf(this.v2ray.flow) >= 0;
-      });
-    }
-  },
   mounted() {
     if (this.which !== null) {
       this.$axios({
         url: apiRoot + "/sharingAddress",
         method: "get",
         params: {
-          touch: this.which
-        }
-      }).then(res => {
+          touch: this.which,
+        },
+      }).then((res) => {
         handleResponse(res, this, () => {
           if (
             res.data.data.sharingAddress.toLowerCase().startsWith("vmess://") ||
@@ -885,25 +708,36 @@ export default {
             this.trojan = this.resolveURL(res.data.data.sharingAddress);
             this.tabChoice = 3;
           } else if (
+            res.data.data.sharingAddress.toLowerCase().startsWith("juicity://")
+          ) {
+            this.juicity = this.resolveURL(res.data.data.sharingAddress);
+            this.tabChoice = 4;
+          } else if (
+            res.data.data.sharingAddress.toLowerCase().startsWith("tuic://")
+          ) {
+            this.tuic = this.resolveURL(res.data.data.sharingAddress);
+            this.tabChoice = 5;
+          } else if (
             res.data.data.sharingAddress.toLowerCase().startsWith("http://") ||
             res.data.data.sharingAddress.toLowerCase().startsWith("https://")
           ) {
             this.http = this.resolveURL(res.data.data.sharingAddress);
-            this.tabChoice = 4;
+            this.tabChoice = 6;
           } else if (
             res.data.data.sharingAddress.toLowerCase().startsWith("socks5://")
           ) {
             this.socks5 = this.resolveURL(res.data.data.sharingAddress);
-            this.tabChoice = 5;
+            this.tabChoice = 7;
           }
           this.$nextTick(() => {
             if (this.readonly) {
               this.$refs.section
                 .querySelectorAll("input, textarea")
-                .forEach(x => (x.readOnly = "readOnly"));
-              this.$refs.section.querySelectorAll("select").forEach(x => {
-                const text = x.querySelector(`option[value="${x.value}"]`)
-                  .textContent;
+                .forEach((x) => (x.readOnly = "readOnly"));
+              this.$refs.section.querySelectorAll("select").forEach((x) => {
+                const text = x.querySelector(
+                  `option[value="${x.value}"]`
+                ).textContent;
                 console.log(x.value, text);
                 x.outerHTML = `<input type="text" class="input" readonly="readonly" value="${text}">`;
               });
@@ -914,7 +748,10 @@ export default {
     }
   },
   methods: {
-    handleV2rayProtocolSwitch() {},
+    variant() {
+      return localStorage["variant"]?.toLowerCase() || "v2ray";
+    },
+    handleV2rayProtocolSwitch() { },
     resolveURL(url) {
       if (url.toLowerCase().startsWith("vmess://")) {
         let obj = JSON.parse(
@@ -934,14 +771,20 @@ export default {
           add: u.host,
           port: u.port,
           id: decodeURIComponent(u.username),
+          flow: u.params.flow || "",
           net: u.params.type || "tcp",
           type: u.params.headerType || "none",
-          host: u.params.sni || u.params.host || "",
+          host: u.params.host || "",
           path: u.params.path || u.params.serviceName || "",
           alpn: u.params.alpn || "",
+          sni: u.params.sni || "",
           tls: u.params.security || "none",
+          fp: u.params.fp || "",
+          pbk: u.params.pbk || "",
+          sid: u.params.sid || "",
+          spx: u.params.spx || "",
           allowInsecure: u.params.allowInsecure || false,
-          protocol: "vless"
+          protocol: "vless",
         };
         if (o.alpn !== "") {
           o.alpn = decodeURIComponent(o.alpn);
@@ -979,7 +822,7 @@ export default {
           obfs: "http",
           plugin: "",
           protocol: "ss",
-          impl: ""
+          impl: "",
         };
         if (u.params.plugin) {
           u.params.plugin = decodeURIComponent(u.params.plugin);
@@ -1049,7 +892,7 @@ export default {
           protoParam: m["protoparam"],
           obfs: pre[4],
           obfsParam: m["obfsparam"],
-          protocol: "ssr"
+          protocol: "ssr",
         };
       } else if (
         url.toLowerCase().startsWith("trojan://") ||
@@ -1063,11 +906,11 @@ export default {
           name: decodeURIComponent(u.hash),
           peer: u.params.peer || u.params.sni || "",
           allowInsecure:
-            u.params.allowInsecure === true || u.params.allowInsecure === "1",
+            u.params.allowInsecure === 'true' || u.params.allowInsecure === "1",
           method: "origin",
           obfs: "none",
           ssCipher: "aes-128-gcm",
-          protocol: "trojan"
+          protocol: "trojan",
         };
         if (url.toLowerCase().startsWith("" + "")) {
           console.log(u.params.encryption);
@@ -1080,7 +923,7 @@ export default {
           const obfsMap = {
             original: "none",
             "": "none",
-            ws: "websocket"
+            ws: "websocket",
           };
           o.obfs = obfsMap[u.params.type || ""];
           if (o.obfs === "ws") {
@@ -1092,6 +935,39 @@ export default {
           }
         }
         return o;
+      } else if (url.toLowerCase().startsWith("juicity://")) {
+        let u = parseURL(url);
+        return {
+          name: decodeURIComponent(u.hash),
+          uuid: decodeURIComponent(u.username),
+          password: decodeURIComponent(u.password),
+          server: u.host,
+          port: u.port,
+          sni: u.params.sni || "",
+          allowInsecure:
+            u.params.allow_insecure === 'true' || u.params.allow_insecure === "1",
+          pinnedCertchainSha256: u.params.pinned_certchain_sha256 || "",
+          cc: u.params.congestion_control || "bbr",
+          protocol: "juicity",
+        };
+      } else if (url.toLowerCase().startsWith("tuic://")) {
+        let u = parseURL(url);
+        return {
+          name: decodeURIComponent(u.hash),
+          uuid: decodeURIComponent(u.username),
+          password: decodeURIComponent(u.password),
+          server: u.host,
+          port: u.port,
+          sni: u.params.sni || "",
+          allowInsecure:
+            u.params.allow_insecure === 'true' || u.params.allow_insecure === "1",
+          disableSni:
+            u.params.disable_sni === 'true' || u.params.disable_sni === "1",
+          alpn: u.params.alpn,
+          cc: u.params.congestion_control || "bbr",
+          udpRelayMode: u.params.udp_relay_mode || "native",
+          protocol: "tuic",
+        };
       } else if (
         url.toLowerCase().startsWith("http://") ||
         url.toLowerCase().startsWith("https://")
@@ -1103,7 +979,7 @@ export default {
           host: u.host,
           port: u.port,
           protocol: u.protocol,
-          name: decodeURIComponent(u.hash)
+          name: decodeURIComponent(u.hash),
         };
       } else if (url.toLowerCase().startsWith("socks5://")) {
         let u = parseURL(url);
@@ -1113,7 +989,7 @@ export default {
           host: u.host,
           port: u.port,
           protocol: u.protocol,
-          name: decodeURIComponent(u.hash)
+          name: decodeURIComponent(u.hash),
         };
       }
       return null;
@@ -1127,13 +1003,14 @@ export default {
           // https://github.com/XTLS/Xray-core/discussions/716
           query = {
             type: srcObj.net,
+            flow: srcObj.flow || "",
             security: srcObj.tls,
+            fp: srcObj.fp || "",
             path: srcObj.path,
             host: srcObj.host,
             headerType: srcObj.type,
-            sni: srcObj.host,
-            flow: srcObj.flow,
-            allowInsecure: srcObj.allowInsecure
+            sni: srcObj.sni,
+            allowInsecure: srcObj.allowInsecure,
           };
           if (srcObj.alpn !== "") {
             query.alpn = srcObj.alpn;
@@ -1144,13 +1021,18 @@ export default {
           if (srcObj.net === "mkcp" || srcObj.net === "kcp") {
             query.seed = srcObj.path;
           }
+          if (query.security == "reality") {
+            query.pbk = srcObj.pbk;
+            query.sid = srcObj.sid;
+            query.spx = srcObj.spx;
+          }
           return generateURL({
             protocol: "vless",
             username: srcObj.id,
             host: srcObj.add,
             port: srcObj.port,
             hash: srcObj.ps,
-            params: query
+            params: query,
           });
         case "vmess":
           //https://github.com/2dust/v2rayN/wiki/%E5%88%86%E4%BA%AB%E9%93%BE%E6%8E%A5%E6%A0%BC%E5%BC%8F%E8%AF%B4%E6%98%8E(ver-2)
@@ -1178,15 +1060,11 @@ export default {
               }
               obj.path = "";
           }
-          if (!(obj.protocol === "vless" && obj.tls === "xtls")) {
-            delete obj.flow;
-          }
           return "vmess://" + Base64.encode(JSON.stringify(obj));
         case "ss":
           /* ss://BASE64(method:password)@server:port#name */
-          tmp = `ss://${Base64.encode(`${srcObj.method}:${srcObj.password}`)}@${
-            srcObj.server
-          }:${srcObj.port}/`;
+          tmp = `ss://${Base64.encode(`${srcObj.method}:${srcObj.password}`)}@${srcObj.server
+            }:${srcObj.port}/`;
           if (srcObj.plugin) {
             const plugin = [srcObj.plugin];
             if (srcObj.plugin === "v2ray-plugin") {
@@ -1228,8 +1106,7 @@ export default {
         case "ssr":
           /* ssr://server:port:proto:method:obfs:URLBASE64(password)/?remarks=URLBASE64(remarks)&protoparam=URLBASE64(protoparam)&obfsparam=URLBASE64(obfsparam)) */
           return `ssr://${Base64.encode(
-            `${srcObj.server}:${srcObj.port}:${srcObj.proto}:${srcObj.method}:${
-              srcObj.obfs
+            `${srcObj.server}:${srcObj.port}:${srcObj.proto}:${srcObj.method}:${srcObj.obfs
             }:${Base64.encodeURI(srcObj.password)}/?remarks=${Base64.encodeURI(
               srcObj.name
             )}&protoparam=${Base64.encodeURI(
@@ -1239,7 +1116,7 @@ export default {
         case "trojan":
           /* trojan://password@server:port?allowInsecure=1&sni=sni#URIESCAPE(name) */
           query = {
-            allowInsecure: srcObj.allowInsecure
+            allowInsecure: srcObj.allowInsecure,
           };
           if (srcObj.peer !== "") {
             query.sni = srcObj.peer;
@@ -1263,7 +1140,47 @@ export default {
             host: srcObj.server,
             port: srcObj.port,
             hash: srcObj.name,
-            params: query
+            params: query,
+          });
+        case "juicity":
+          query = {
+            allow_insecure: srcObj.allowInsecure,
+            congestion_control: srcObj.cc,
+          };
+          if (srcObj.sni !== "") {
+            query.sni = srcObj.sni;
+          }
+          if (srcObj.pinnedCertchainSha256 !== "") {
+            query.pinned_certchain_sha256 = srcObj.pinnedCertchainSha256;
+          }
+          return generateURL({
+            protocol: "juicity",
+            username: srcObj.uuid,
+            password: srcObj.password,
+            host: srcObj.server,
+            port: srcObj.port,
+            hash: srcObj.name,
+            params: query,
+          });
+        case "tuic":
+          query = {
+            allow_insecure: srcObj.allowInsecure,
+            congestion_control: srcObj.cc,
+            disable_sni: srcObj.disableSni,
+            alpn: srcObj.alpn,
+            udp_relay_mode: srcObj.udpRelayMode,
+          };
+          if (srcObj.sni !== "") {
+            query.sni = srcObj.sni;
+          }
+          return generateURL({
+            protocol: "tuic",
+            username: srcObj.uuid,
+            password: srcObj.password,
+            host: srcObj.server,
+            port: srcObj.port,
+            hash: srcObj.name,
+            params: query,
           });
         case "http":
         case "https":
@@ -1271,12 +1188,12 @@ export default {
             protocol: srcObj.protocol + "-proxy",
             host: srcObj.host,
             port: srcObj.port,
-            hash: srcObj.name
+            hash: srcObj.name,
           };
           if (srcObj.username && srcObj.password) {
             Object.assign(tmp, {
               username: srcObj.username,
-              password: srcObj.password
+              password: srcObj.password,
             });
           }
           return generateURL(tmp);
@@ -1285,12 +1202,12 @@ export default {
             protocol: "socks5",
             host: srcObj.host,
             port: srcObj.port,
-            hash: srcObj.name
+            hash: srcObj.name,
           };
           if (srcObj.username && srcObj.password) {
             Object.assign(tmp, {
               username: srcObj.username,
-              password: srcObj.password
+              password: srcObj.password,
             });
           }
           return generateURL(tmp);
@@ -1305,7 +1222,7 @@ export default {
           type: "is-warning",
           position: "is-top",
           queue: false,
-          duration: 5000
+          duration: 5000,
         });
         this.$nextTick(() => {
           this.v2ray.tls = "tls";
@@ -1330,10 +1247,16 @@ export default {
         if (this.tabChoice === 3 && !k.startsWith("trojan_")) {
           continue;
         }
-        if (this.tabChoice === 4 && !k.startsWith("http_")) {
+        if (this.tabChoice === 4 && !k.startsWith("juicity_")) {
           continue;
         }
-        if (this.tabChoice === 5 && !k.startsWith("socks5_")) {
+        if (this.tabChoice === 5 && !k.startsWith("tuic_")) {
+          continue;
+        }
+        if (this.tabChoice === 6 && !k.startsWith("http_")) {
+          continue;
+        }
+        if (this.tabChoice === 7 && !k.startsWith("socks5_")) {
           continue;
         }
         let x = this.$refs[k];
@@ -1367,7 +1290,7 @@ export default {
             type: "is-danger",
             hasIcon: true,
             onConfirm: () => true,
-            onCancel: () => false
+            onCancel: () => false,
           });
           if (!result) {
             return;
@@ -1381,23 +1304,23 @@ export default {
       } else if (this.tabChoice === 3) {
         coded = this.generateURL(this.trojan);
       } else if (this.tabChoice === 4) {
-        coded = this.generateURL(this.http);
+        coded = this.generateURL(this.juicity);
       } else if (this.tabChoice === 5) {
+        coded = this.generateURL(this.tuic);
+      } else if (this.tabChoice === 6) {
+        coded = this.generateURL(this.http);
+      } else if (this.tabChoice === 7) {
         coded = this.generateURL(this.socks5);
       }
       this.$emit("submit", coded);
-    }
-  }
+    },
+  },
 };
 </script>
 
 <style lang="scss">
 .is-twitter .is-active a {
   color: #4099ff !important;
-}
-
-.readonly {
-  pointer-events: none;
 }
 
 .same-width-5 li {
