@@ -17,7 +17,7 @@ export default {
     none: "none",
     optional: "optional",
     loadBalance: "Load Balance",
-    log: "Logs"
+    log: "Logs",
   },
   welcome: {
     title: "Welcome",
@@ -26,12 +26,12 @@ export default {
     newVersion: "Detected new version: {version}",
     messages: [
       "There is no server.",
-      "You can create/import a server or import a subscription."
-    ]
+      "You can create/import a server or import a subscription.",
+    ],
   },
   v2ray: {
     start: "Start",
-    stop: "Stop"
+    stop: "Stop",
   },
   server: {
     name: "Server Name",
@@ -44,21 +44,21 @@ export default {
       notAllowInsecure:
         "According to the docs of {name}, if you use {name}, AllowInsecure will be forbidden.",
       notRecommend:
-        "According to the docs of {name}, if you use {name}, AllowInsecure is not recommend."
-    }
+        "According to the docs of {name}, if you use {name}, AllowInsecure is not recommend.",
+    },
   },
   InSecureConfirm: {
     title: "Dangerous configuration detected",
     message:
       "The configuration has set the <b>AllowInsecure</b> to true. This may cause security risks. Are you sure to continue?",
     confirm: "I know what I'm doing",
-    cancel: "cancel"
+    cancel: "cancel",
   },
   subscription: {
     host: "Host",
     remarks: "Remarks",
     timeLastUpdate: "Datetime of Last Update",
-    numberServers: "Number of Servers"
+    numberServers: "Number of Servers",
   },
   operations: {
     name: "Operations",
@@ -86,20 +86,20 @@ export default {
     yes: "Yes",
     no: "No",
     switchSite: "Switch to alternate site",
-    addOutbound: "Add an outbound"
+    addOutbound: "Add an outbound",
   },
   register: {
     title: "Create an admin account first",
     messages: [
       "Remember your admin account which is importantly used to login.",
       "Account information is stored in local. We never send information to any server.",
-      "Once password was forgot, you could use v2raya --reset-password to reset."
-    ]
+      "Once password was forgot, you could use v2raya --reset-password to reset.",
+    ],
   },
   login: {
     title: "Login",
     username: "Username",
-    password: "Password"
+    password: "Password",
   },
   setting: {
     transparentProxy: "Transparent Proxy/System Proxy",
@@ -135,7 +135,7 @@ export default {
       dependTransparentMode: "Follows Transparent Proxy/System Proxy",
       closed: "Off",
       advanced: "Advanced Setting",
-      leastPing: "Least Latency First"
+      leastPing: "Least Latency First",
     },
     messages: {
       gfwlist:
@@ -163,8 +163,8 @@ export default {
       ssPluginImpl:
         "★default: 'transport' for simple-obfs, 'chained' for v2ray-plugin." +
         "★chained: shadowsocks traffic will be redirect to standalone plugin." +
-        "★transport: processed by the transport layer of v2ray/xray core directly."
-    }
+        "★transport: processed by the transport layer of v2ray/xray core directly.",
+    },
   },
   customAddressPort: {
     title: "Address and Ports",
@@ -179,8 +179,8 @@ export default {
       "Service address default as 0.0.0.0:2017 can be changed by setting environment variable <code>V2RAYA_ADDRESS</code> and command argument<code>--address</code>.",
       "If you start v2raya docker container with port mapping instead of <code>--network host</code>, you can remapping ports in this way.",
       "We cannot judge port occupations in docker mode. Confirm it by yourself.",
-      "Zero means to close this port."
-    ]
+      "Zero means to close this port.",
+    ],
   },
   customRouting: {
     title: "Customize Routing Rule",
@@ -198,8 +198,8 @@ export default {
       1: 'To make a SiteDat file by yourself: <a href="https://github.com/ToutyRater/V2Ray-SiteDAT">ToutyRater/V2Ray-SiteDAT</a>',
       2: "Multi-select is supported.",
       noSiteDatFileFound: "No siteDat file found in {V2RayLocationAsset}",
-      emptyRuleNotPermitted: "Empty rule is not permitted"
-    }
+      emptyRuleNotPermitted: "Empty rule is not permitted",
+    },
   },
   doh: {
     title: "Configure DoH Server",
@@ -211,8 +211,8 @@ export default {
       "USA: cloudflare, dns.google, etc",
       'Checklist：<a href="https://dnscrypt.info/public-servers" target="_blank">public-servers</a>',
       'Besides, setting up DoH service at your own native server is suggested and well-behaved in most cases <a href="https://github.com/facebookexperimental/doh-proxy" target="_blank">doh-proxy</a>. In this case, it is recommended to run the server(doh-proxy/doh-httpproxy) providing service and client(doh-stub) connecting to doh.opendns.com at the same time and connect them in series, because you can hardly find a server that is not polluted in a generally contaminated region.',
-      "Optimally, place one or two lines above. The list will restore to default after saving with empty content."
-    ]
+      "Optimally, place one or two lines above. The list will restore to default after saving with empty content.",
+    ],
   },
   dns: {
     title: "Configure DNS Server",
@@ -220,8 +220,8 @@ export default {
     externalQueryServers: "External Domain Query Servers",
     messages: [
       '"@:(dns.internalQueryServers)" are designed to be used to look up domain names in China, while "@:(dns.externalQueryServers)" be used to look up others.',
-      '"@:(dns.internalQueryServers)" will be used to look up all domain names if "@:(dns.externalQueryServers)" is empty.'
-    ]
+      '"@:(dns.internalQueryServers)" will be used to look up all domain names if "@:(dns.externalQueryServers)" is empty.',
+    ],
   },
   egressPortWhitelist: {
     title: "Egress Port Whitelist",
@@ -232,8 +232,8 @@ export default {
       "Transparent proxy will force all TCP and UDP traffic to pass through proxy server B, where source IP address will be replaced with proxy B's. Moreover, if some clients send requests to server A that provides service, they will received responses from your proxy B's IP address weirdly, which is illegal.",
       "To resolve it, we need to add those service ports to whitelist so that not pass through proxy.For examples, ssh(22)、v2raya({v2rayaPort}).",
       "Obviously, if the server does not provide any service, you can skip configuring.",
-      "Formatting：22 means port 22，20170:20172 means three ports 20170 to 20172."
-    ]
+      "Formatting：22 means port 22，20170:20172 means three ports 20170 to 20172.",
+    ],
   },
   configureServer: {
     title: "Configure Server | Server",
@@ -252,31 +252,33 @@ export default {
     seedObfuscation: "Seed",
     username: "Username",
     password: "Password",
-    origin: "origin"
+    origin: "origin",
+    pinnedCertchainSha256: "pinned certificate chain sha256"
   },
   configureSubscription: {
-    title: "Configure Subscription"
+    title: "Configure Subscription",
   },
   import: {
     message: "Input a server link or subscription address:",
     batchMessage: "One server link per line:",
-    qrcodeError: "Failed to find a valid QRCode, please try again"
+    qrcodeError: "Failed to find a valid QRCode, please try again",
   },
   delete: {
     title: "Confirm to DELETE",
     message:
-      "Be sure to <b>DELETE</b> those servers/subscriptions? It is not reversible."
+      "Be sure to <b>DELETE</b> those servers/subscriptions? It is not reversible.",
   },
   latency: {
     message:
-      "Latency tests used to cost one or several minutes. Wait patiently please."
+      "Latency tests used to cost one or several minutes. Wait patiently please.",
   },
   version: {
     higherVersionNeeded:
       "This operation need higher version of v2rayA than {version}",
-    v2rayInvalid: "geosite.dat, geoip.dat or v2ray-core may not be installed correctly",
+    v2rayInvalid:
+      "geosite.dat, geoip.dat or v2ray-core may not be installed correctly",
     v2rayNotV5:
-      "The version of v2ray-core is not v5. Use v5 or downgrade v2rayA to v1.5"
+      "The version of v2ray-core is not v5. Use v5 or downgrade v2rayA to v1.5",
   },
   about: `<p>v2rayA is a web GUI client of V2Ray.</p>
           <p class="about-small">Default ports:</p>
@@ -296,24 +298,25 @@ export default {
         "Cannot find v2rayA at {url}. Make sure v2rayA is running at this address.",
       cannotCommunicate: [
         "Cannot communicate. If your service is running and ports open correctly, the reason may be that current browser does not allow https sites to access http resources, you can try using Chrome or switching to alternate http site.",
-        "Cannot communicate. Firefox does not allow https sites to access http resources, you can try switching to alternate http sites."
-      ]
+        "Cannot communicate. Firefox does not allow https sites to access http resources, you can try switching to alternate http sites.",
+      ],
     },
     urls: {
-      usage: "https://github.com/v2rayA/v2rayA/wiki/Usage"
-    }
+      usage: "https://github.com/v2rayA/v2rayA/wiki/Usage",
+    },
   },
   routingA: {
-    messages: ["click the button 'Help&Manual' for help"]
+    messages: ["click the button 'Help&Manual' for help"],
   },
   outbound: {
     addMessage: "Please input the outbound name you want to add:",
     deleteMessage:
-      'Be sure to <b>DELETE</b> the outbound "{outboundName}"? It is not reversible.'
+      'Be sure to <b>DELETE</b> the outbound "{outboundName}"? It is not reversible.',
   },
   log: {
     logModalTitle: "View logs",
     refreshInterval: "Refresh Interval",
-    seconds: "seconds"
-  }
+    seconds: "seconds",
+    autoScoll: "Auto Scroll",
+  },
 };
