@@ -1,7 +1,9 @@
 import zh from "./zh";
 import en from "./en";
+import fa from "./fa-ir";
 
 export default {
   zh,
-  en
+  en,
+  fa
 };
