@@ -25,7 +25,7 @@ Vue.prototype.$remount = () => {
   vue = new Vue({
     i18n,
     store,
-    render: h => h(App)
+    render: (h) => h(App),
   }).$mount("#app");
 };
 
