@@ -2,31 +2,25 @@ var webpack = require("webpack");
 var path = require("path");
 var WebpackIconfontPluginNodejs = require("webpack-iconfont-plugin-nodejs");
 var dir = "src/assets/iconfont";
-var UglifyJsPlugin = require("uglifyjs-webpack-plugin");
+var TerserPlugin = require("terser-webpack-plugin");
 
 module.exports = {
-  configureWebpack: config => {
+  configureWebpack: (config) => {
     config.resolve.alias["vue$"] = "vue/dist/vue.esm.js";
     return {
       optimization: {
         minimizer: [
-          new UglifyJsPlugin({
-            uglifyOptions: {
-              output: {
-                beautify: false
-              },
-              compress: {
-                warnings: false,
-                drop_console: true
-              },
-              sourceMap: false
-            }
-          })
-        ]
+          new TerserPlugin({
+            minify: TerserPlugin.swcMinify,
+            terserOptions: {
+              compress: true,
+            },
+          }),
+        ],
       },
       plugins: [
         new webpack.DefinePlugin({
-          apiRoot: '`${localStorage["backendAddress"]}/api`'
+          apiRoot: '`${localStorage["backendAddress"]}/api`',
         }),
         new WebpackIconfontPluginNodejs({
           cssPrefix: "icon",
@@ -35,17 +29,17 @@ module.exports = {
           fontsOutput: path.join(dir, "fonts/"),
           cssOutput: path.join(dir, "fonts/font.css"),
           // htmlOutput: path.join(dir, "fonts/_font-preview.html"),
-          jsOutput: path.join(dir, "fonts/fonts.js")
+          jsOutput: path.join(dir, "fonts/fonts.js"),
           // formats: ['ttf', 'woff', 'svg'],
-        })
-      ]
+        }),
+      ],
     };
   },
 
   productionSourceMap: false,
 
   devServer: {
-    port: 8081
+    port: 8081,
   },
 
   // publicPath: "./static/",
@@ -67,5 +61,5 @@ module.exports = {
 
   lintOnSave: false,
 
-  transpileDependencies: ["buefy"]
+  transpileDependencies: ["buefy"],
 };
