@@ -15,7 +15,7 @@ iptables -w 2 -A DROP_SPOOFING -p udp --sport 53 -m string --algo bm --hex-strin
 iptables -w 2 -A DROP_SPOOFING -p udp --sport 53 -m string --algo bm --hex-string "|0010fc|" --from 60 --to 180 -j DROP
 iptables -w 2 -A DROP_SPOOFING -p udp --sport 53 -m string --algo bm --hex-string "|0010fd|" --from 60 --to 180 -j DROP
 iptables -w 2 -A DROP_SPOOFING -p udp --sport 53 -m string --algo bm --hex-string "|0010ff|" --from 60 --to 180 -j DROP
-iptables -w 2 -I INPUT -j DROP_SPOOFING 
+iptables -w 2 -I INPUT -j DROP_SPOOFING
 iptables -w 2 -I FORWARD -j DROP_SPOOFING
 `
 	if IsIPv6Supported() {
@@ -34,7 +34,7 @@ ip6tables -w 2 -I FORWARD -j DROP_SPOOFING
 `
 	}
 	return Setter{
-		Cmds:      commands,
+		Cmds: commands,
 	}
 }
 
@@ -47,13 +47,13 @@ iptables -w 2 -X DROP_SPOOFING
 `
 	if IsIPv6Supported() {
 		commands += `
-ip6tables -w 2 -D INPUT -j DROP_SPOOFING 
-ip6tables -w 2 -D FORWARD -j DROP_SPOOFING 
+ip6tables -w 2 -D INPUT -j DROP_SPOOFING
+ip6tables -w 2 -D FORWARD -j DROP_SPOOFING
 ip6tables -w 2 -F DROP_SPOOFING
 ip6tables -w 2 -X DROP_SPOOFING
 `
 	}
 	return Setter{
-		Cmds:      commands,
+		Cmds: commands,
 	}
 }
