@@ -31,11 +31,11 @@ func NewSOCKS(link string) (ServerObj, error) {
 func ParseSocksURL(u string) (data *SOCKS, err error) {
 	t, err := url.Parse(u)
 	if err != nil {
-		return nil, InvalidParameterErr
+		return nil, ErrInvalidParameter
 	}
 	port, err := strconv.Atoi(t.Port())
 	if err != nil {
-		return nil, InvalidParameterErr
+		return nil, ErrInvalidParameter
 	}
 	data = &SOCKS{
 		Name:   t.Fragment,
