@@ -3,6 +3,13 @@ package serverObj
 import (
 	"encoding/base64"
 	"fmt"
+	"net"
+	"net/url"
+	"regexp"
+	"strconv"
+	"strings"
+	"time"
+
 	jsoniter "github.com/json-iterator/go"
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
@@ -10,12 +17,6 @@ import (
 	"github.com/v2rayA/v2rayA/common/ntp"
 	"github.com/v2rayA/v2rayA/core/coreObj"
 	"github.com/v2rayA/v2rayA/pkg/util/log"
-	"net"
-	"net/url"
-	"regexp"
-	"strconv"
-	"strings"
-	"time"
 )
 
 func init() {
@@ -39,9 +40,13 @@ type V2Ray struct {
 	Net           string `json:"net"`
 	Type          string `json:"type"`
 	Host          string `json:"host"`
-	SNI           string `json:"sni"`
+	SNI           string `json:"sni,omitempty"`
 	Path          string `json:"path"`
 	TLS           string `json:"tls"`
+	Fingerprint   string `json:"fingerprint,omitempty"`
+	PublicKey     string `json:"pbk,omitempty"`
+	ShortId       string `json:"sid,omitempty"`
+	SpiderX       string `json:"spx,omitempty"`
 	Flow          string `json:"flow,omitempty"`
 	Alpn          string `json:"alpn,omitempty"`
 	AllowInsecure bool   `json:"allowInsecure"`
@@ -55,7 +60,7 @@ func NewV2Ray(link string) (ServerObj, error) {
 	} else if strings.HasPrefix(link, "vless://") {
 		return ParseVlessURL(link)
 	}
-	return nil, InvalidParameterErr
+	return nil, ErrInvalidParameter
 }
 
 func ParseVlessURL(vless string) (data *V2Ray, err error) {
@@ -68,15 +73,21 @@ func ParseVlessURL(vless string) (data *V2Ray, err error) {
 		Add:           u.Hostname(),
 		Port:          u.Port(),
 		ID:            u.User.String(),
+		Aid:           u.Query().Get("aid"),
 		Net:           u.Query().Get("type"),
 		Type:          u.Query().Get("headerType"),
 		Host:          u.Query().Get("host"),
 		SNI:           u.Query().Get("sni"),
 		Path:          u.Query().Get("path"),
 		TLS:           u.Query().Get("security"),
+		Fingerprint:   u.Query().Get("fp"),
+		PublicKey:     u.Query().Get("pbk"),
+		ShortId:       u.Query().Get("sid"),
+		SpiderX:       u.Query().Get("spx"),
 		Flow:          u.Query().Get("flow"),
 		Alpn:          u.Query().Get("alpn"),
 		AllowInsecure: u.Query().Get("allowInsecure") == "true",
+		V:             vless,
 		Protocol:      "vless",
 	}
 	if data.Net == "" {
@@ -88,11 +99,11 @@ func ParseVlessURL(vless string) (data *V2Ray, err error) {
 	if data.Type == "" {
 		data.Type = "none"
 	}
+	if data.Host == "" {
+		data.Host = u.Query().Get("host")
+	}
 	if data.TLS == "" {
 		data.TLS = "none"
-	}
-	if data.Flow == "" {
-		data.Flow = "xtls-rprx-direct"
 	}
 	if data.Net == "mkcp" || data.Net == "kcp" {
 		data.Path = u.Query().Get("seed")
@@ -118,7 +129,7 @@ func ParseVmessURL(vmess string) (data *V2Ray, err error) {
 		s := strings.Split(vmess[8:], "?")[0]
 		s, err = common.Base64StdDecode(s)
 		if err != nil {
-			s, err = common.Base64URLDecode(s)
+			s, _ = common.Base64URLDecode(s)
 		}
 		subMatch := re.FindStringSubmatch(s)
 		if subMatch == nil {
@@ -199,8 +210,12 @@ func (v *V2Ray) Configuration(info PriorInfo) (c Configuration, err error) {
 	switch strings.ToLower(v.Protocol) {
 	case "vmess", "vless":
 		id := v.ID
+		network := v.Net
 		if l := len([]byte(id)); l < 32 || l > 36 {
 			id = common.StringToUUID5(id)
+		}
+		core.StreamSettings = &coreObj.StreamSettings{
+			Network: network,
 		}
 		switch strings.ToLower(v.Protocol) {
 		case "vmess":
@@ -208,11 +223,28 @@ func (v *V2Ray) Configuration(info PriorInfo) (c Configuration, err error) {
 				return Configuration{}, fmt.Errorf("please sync datetime first. Your datetime is %v, and the "+
 					"correct datetime is %v", time.Now().Local().Format(ntp.DisplayFormat), t.Local().Format(ntp.DisplayFormat))
 			}
-			if len(v.Aid) > 0 && v.Aid != "0" {
-				log.Error("We do not support alterID > 0 since v1.5.6 due to security; please contact your server provider or " +
-					"downgrade v2rayA to v1.5.5")
-				return Configuration{}, fmt.Errorf("unsupported alterID > 0: check the log for more information")
+			security := v.Security
+			if security == "" {
+				security = "auto"
 			}
+			var aid int
+			if _aid, err := strconv.Atoi(v.Aid); err == nil {
+				aid = _aid
+			}
+			core.Settings.Vnext = []coreObj.Vnext{
+				{
+					Address: v.Add,
+					Port:    port,
+					Users: []coreObj.User{
+						{
+							ID:       id,
+							AlterID:  aid,
+							Security: security,
+						},
+					},
+				},
+			}
+		case "vless":
 			security := v.Security
 			if security == "" {
 				security = "auto"
@@ -223,30 +255,20 @@ func (v *V2Ray) Configuration(info PriorInfo) (c Configuration, err error) {
 					Port:    port,
 					Users: []coreObj.User{
 						{
-							ID:       id,
-							AlterID:  0,
-							Security: security,
-						},
-					},
-				},
-			}
-		case "vless":
-			core.Settings.Vnext = []coreObj.Vnext{
-				{
-					Address: v.Add,
-					Port:    port,
-					Users: []coreObj.User{
-						{
-							ID: id,
-							//AlterID:    0, // keep AEAD on
+							ID:         id,
 							Encryption: "none",
 						},
 					},
 				},
 			}
-		}
-		core.StreamSettings = &coreObj.StreamSettings{
-			Network: v.Net,
+			// if network == "tcp" {
+			// 	tcpSetting := coreObj.TCPSettings{
+			// 		Header: coreObj.TCPHeader{
+			// 			Type: "none",
+			// 		},
+			// 	}
+			// 	core.StreamSettings.TCPSettings = &tcpSetting
+			// }
 		}
 		// 根据传输协议(network)修改streamSettings
 		//TODO: QUIC
@@ -333,6 +355,8 @@ func (v *V2Ray) Configuration(info PriorInfo) (c Configuration, err error) {
 					Path: v.Path,
 				}
 			}
+		default:
+			return Configuration{}, fmt.Errorf("unexpected transport type: %v", v.Net)
 		}
 		if strings.ToLower(v.TLS) == "tls" {
 			core.StreamSettings.Security = "tls"
@@ -354,20 +378,19 @@ func (v *V2Ray) Configuration(info PriorInfo) (c Configuration, err error) {
 				}
 				core.StreamSettings.TLSSettings.Alpn = alpn
 			}
+			// uTLS fingerprint
+			core.StreamSettings.TLSSettings.Fingerprint = v.Fingerprint
 		} else if strings.ToLower(v.TLS) == "xtls" {
 			core.StreamSettings.Security = "xtls"
 			core.StreamSettings.XTLSSettings = &coreObj.TLSSettings{}
 			// SNI
-			if v.Host != "" {
-				core.StreamSettings.XTLSSettings.ServerName = v.Host
+			if v.SNI != "" {
+				core.StreamSettings.XTLSSettings.ServerName = v.SNI
 			} else if v.Host != "" {
-				core.StreamSettings.TLSSettings.ServerName = v.Host
+				core.StreamSettings.XTLSSettings.ServerName = v.Host
 			}
 			if v.AllowInsecure {
 				core.StreamSettings.XTLSSettings.AllowInsecure = true
-			}
-			if v.Flow == "" {
-				v.Flow = "xtls-rprx-origin"
 			}
 			if v.Alpn != "" {
 				alpn := strings.Split(v.Alpn, ",")
@@ -376,6 +399,19 @@ func (v *V2Ray) Configuration(info PriorInfo) (c Configuration, err error) {
 				}
 				core.StreamSettings.XTLSSettings.Alpn = alpn
 			}
+		} else if strings.ToLower(v.TLS) == "reality" {
+			core.StreamSettings.Security = "reality"
+			core.StreamSettings.RealitySettings = &coreObj.RealitySettings{
+				ServerName:  v.SNI,
+				Fingerprint: v.Fingerprint,
+				Show:        false,
+				PublicKey:   v.PublicKey,
+				ShortID:     v.ShortId,
+				SpiderX:     v.SpiderX,
+			}
+		}
+		// Flow
+		if v.Flow != "" {
 			vnext := core.Settings.Vnext.([]coreObj.Vnext)
 			vnext[0].Users[0].Flow = v.Flow
 			core.Settings.Vnext = vnext
@@ -409,14 +445,17 @@ func (v *V2Ray) ExportToURL() string {
 		case "grpc":
 			setValue(&query, "serviceName", v.Path)
 		}
-		//TODO: QUIC
 		if v.TLS != "none" {
-			setValue(&query, "sni", v.Host) // FIXME: it may be different from ws's host
+			setValue(&query, "flow", v.Flow)
+			setValue(&query, "sni", v.SNI)
 			setValue(&query, "alpn", v.Alpn)
 			setValue(&query, "allowInsecure", strconv.FormatBool(v.AllowInsecure))
-		}
-		if v.TLS == "xtls" {
-			setValue(&query, "flow", v.Flow)
+			setValue(&query, "fp", v.Fingerprint)
+			if v.TLS == "reality" {
+				setValue(&query, "pbk", v.PublicKey)
+				setValue(&query, "sid", v.ShortId)
+				setValue(&query, "spx", v.SpiderX)
+			}
 		}
 
 		U := url.URL{
