@@ -12,6 +12,7 @@ import (
 	"net"
 	"net/http"
 	"strings"
+	"sync"
 )
 
 // HTTPObfs is shadowsocks http simple-obfs implementation
@@ -24,9 +25,13 @@ type HTTPObfs struct {
 	offset        int
 	firstRequest  bool
 	firstResponse bool
+	rMu           sync.Mutex
+	wMu           sync.Mutex
 }
 
 func (ho *HTTPObfs) Read(b []byte) (int, error) {
+	ho.rMu.Lock()
+	defer ho.rMu.Unlock()
 	if ho.buf != nil {
 		n := copy(b, ho.buf[ho.offset:])
 		ho.offset += n
@@ -64,6 +69,8 @@ func (ho *HTTPObfs) Read(b []byte) (int, error) {
 }
 
 func (ho *HTTPObfs) Write(b []byte) (int, error) {
+	ho.wMu.Lock()
+	defer ho.wMu.Unlock()
 	if ho.firstRequest {
 		randBytes := make([]byte, 16)
 		rand.Read(randBytes)
@@ -71,7 +78,6 @@ func (ho *HTTPObfs) Write(b []byte) (int, error) {
 		req.Header.Set("User-Agent", fmt.Sprintf("curl/7.%d.%d", rand.Int()%54, rand.Int()%2))
 		req.Header.Set("Upgrade", "websocket")
 		req.Header.Set("Connection", "Upgrade")
-		req.Host = ho.host
 		if ho.port != "80" {
 			req.Host = fmt.Sprintf("%s:%s", ho.host, ho.port)
 		}
