@@ -2,10 +2,11 @@ package service
 
 import (
 	"fmt"
-	"github.com/v2rayA/v2rayA/conf"
 	url2 "net/url"
 	"strings"
 	"time"
+
+	"github.com/v2rayA/v2rayA/conf"
 
 	"github.com/v2rayA/v2rayA/common"
 	"github.com/v2rayA/v2rayA/common/httpClient"
@@ -39,8 +40,11 @@ func Import(url string, which *configure.Which) (err error) {
 		}
 		return err
 	}
-	supportedPrefix := []string{"vmess://", "vless://", "ss://", "ssr://", "trojan://", "trojan-go://", "http-proxy://",
-		"https-proxy://", "socks5://", "http2://"}
+	supportedPrefix := []string{"vmess", "vless", "ss", "ssr", "trojan", "trojan-go", "http-proxy",
+		"https-proxy", "socks5", "http2", "juicity", "tuic"}
+	for i := range supportedPrefix {
+		supportedPrefix[i] += "://"
+	}
 	if PluginManagerValidateLink(url) || common.HasAnyPrefix(url, supportedPrefix) {
 		var obj serverObj.ServerObj
 		obj, err = ResolveURL(url)
