@@ -3,6 +3,14 @@ package service
 import (
 	"errors"
 	"fmt"
+	"io"
+	"net"
+	"net/http"
+	"strconv"
+	"strings"
+	"sync"
+	"time"
+
 	"github.com/v2rayA/v2rayA/common/httpClient"
 	"github.com/v2rayA/v2rayA/common/resolv"
 	"github.com/v2rayA/v2rayA/core/coreObj"
@@ -10,12 +18,6 @@ import (
 	"github.com/v2rayA/v2rayA/core/v2ray"
 	"github.com/v2rayA/v2rayA/db/configure"
 	"github.com/v2rayA/v2rayA/pkg/util/log"
-	"net"
-	"net/http"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 )
 
 const HttpTestURL = "https://gstatic.com/generate_204"
@@ -135,21 +137,33 @@ func TestHttpLatency(which []*configure.Which, timeout time.Duration, maxParalle
 	}
 	inboundPortMap := make([]string, len(vms))
 	pluginPortMap := make(map[int]int)
-	var toClose []net.Listener
+	var toClose []io.Closer
+	defer func() {
+		for _, l := range toClose {
+			_ = l.Close()
+		}
+	}()
 	for i, v := range vms {
 		if which[i].Latency != "" {
 			continue
 		}
 		//find a port for the inbound
+		t := time.Now()
 		var port int
 		for {
 			l, err := net.Listen("tcp", "0.0.0.0:0")
 			if err == nil {
-				toClose = append(toClose, l)
 				port = l.Addr().(*net.TCPAddr).Port
-				break
+				toClose = append(toClose, l)
+				l2, err2 := net.ListenPacket("udp", "0.0.0.0:"+strconv.Itoa(port))
+				if err2 == nil {
+					toClose = append(toClose, l2)
+					break
+				}
 			}
-			time.Sleep(30 * time.Millisecond)
+			if time.Since(t) > 3*time.Second {
+				return nil, fmt.Errorf("timeout: failed to find availble ports")
+			}
 		}
 		v2rayInboundPort := strconv.Itoa(port)
 		pluginPort := 0
@@ -160,9 +174,15 @@ func TestHttpLatency(which []*configure.Which, timeout time.Duration, maxParalle
 				if err == nil {
 					toClose = append(toClose, l)
 					port = l.Addr().(*net.TCPAddr).Port
-					break
+					l2, err2 := net.ListenPacket("udp", "127.0.0.1:"+strconv.Itoa(port))
+					if err2 == nil {
+						toClose = append(toClose, l2)
+						break
+					}
 				}
-				time.Sleep(30 * time.Millisecond)
+				if time.Since(t) > 3*time.Second {
+					return nil, fmt.Errorf("timeout: failed to find availble ports")
+				}
 			}
 			pluginPort = port
 			pluginPortMap[i] = port
@@ -180,11 +200,11 @@ func TestHttpLatency(which []*configure.Which, timeout time.Duration, maxParalle
 	for _, l := range toClose {
 		_ = l.Close()
 	}
+	toClose = nil
 	time.Sleep(30 * time.Millisecond)
 	tmpl.Routing.DomainStrategy = "AsIs"
 	addHosts(tmpl, vms)
 	tmpl.SetOutboundSockopt()
-
 	if err := v2ray.ProcessManager.Start(tmpl); err != nil {
 		return nil, err
 	}
